### 2020-09-22 #46 @pboesu
[Re] Modeling Insect Phenology Using Ordinal Regression and Continuation Ratio Models
**Original articles:**
- B. Dennis, W.P. Kemp,and R.C. Beckwith (1986): "Stochastic Model of Insect Phenology: Estimation and Testing." Environmental Entomology 15(3):540–546. https://doi.org/10.1093/ee/15.3.540
- S.G. Candy (1991): "Modeling Insect Phenology Using Ordinal Regression and Continuation Ratio Models." Environmental Entomology 20(1):190–195. https://doi.org/10.1093/ee/20.1.190

**Scientific domain:** Ecology
**Programming language:** R
**Suggested editor:** @karthik @tpoisot

### Editor (@tpoisot) checklist

- Reviewer 1
    - [x] assigned @juniperlsimonis
    - [x] returned
- Reviewer 2
    - [x] assigned @aammd
    - [x] returned
