+++ article/article.tex
@@ -2,9 +2,11 @@
 
 \usepackage{booktabs}
 \usepackage{afterpage}
+\usepackage{lineno}
 \begin{document}
 
 \input{header.tex}
+\linenumbers
 \input{content.tex}
 
 \hypersetup{linkcolor=black,urlcolor=darkgray}
