### 2021-03-19 #46 @pboesu
@rougier Just to declare, that I am still interested in having this submission peer-reviewed. I do appreciate the difficulties of securing editor and reviewer availability in the current conditions (and am guilty myself of delaying a ReScience submission last year), but it would be great if this manuscript could be progressed in the not too distant future.
