### 2021-03-24 #46 @rougier
Ouups, I'm deeply sorry for this very long delay, for some reason it went out of my radar. We'll try to accelerate things.

@karthik or @tpoisot Do you have time to edit this submission "Insect Phenology Using Ordinal Regression and Continuation Ratio Models / Ecology / R"
