### 2021-03-24 #46 @rougier
@FlukeAndFeather @juniperlsimonis @trashbirdecology @ha0ye @jsta @opetchey @karthik @dmcglinn
Dear all, can any of you review this submission in ecology / R for ReScience C (Modeling Insect Phenology Using Ordinal Regression and Continuation Ratio Models)?
