### 2021-03-24 #46 @tpoisot
I haven't used R in just about forever, so I'm not sure I'm the best person to do this. Maybe @aammd ?
