### 2021-03-29 #46 @rougier
@tpoisot @juniperlsimonis @aammd The review instruction are available at https://rescience.github.io/edit/.
Don't hesitate to poke me if needed.
