### 2021-04-27 pboesu/replication_candy_1991#1 @aammd
Review for Rescience
This is for your submission to Rescience, issue ReScience/submissions#46

I really like this work! I have read through your article twice, run all the code, and quickly read both replicated papers. I also have some slight experience with ordinal models; I've fit one using brms. Also I am very enthusiastic about insects 🐛 🦋 !

In brief, I think this is excellent. Not only have you reproduced everything in the original but you've improved them in almost every aspect. The data are better organized. The math is more clearly explained. The code is modernized and well commented. The figures are identical except for some helpful colour.

There are precious few comments or improvements to offer but here are some that occurred to me:

### The code
* `renv` worked flawlessly in my case.
* I like the numbered scripts, organized with `source`
* great and readable comments
* I am asked to ascertain that "The clarity of the code and instructions must also be judged in the light of the conventions of the field". I think your code and instructions are exemplary and would happily teach from it. (Style-wise, I personally mix base and tidyverse in nearly the same fashion! so according to my personal conventions this is quite correct)

minor points?

I got all the errors which you forwarn in your README, with one extra. When I ran `code/03_fit_dennis_et_al_model.R` I got
`1: In sqrt(BB * ddeg[i]) : NaNs produced`

clearly `optim` is feeding in some negative values for `BB`. I wonder if it would work to stick in into a log link, i.e.
`sqrt(exp(BB_log) * ddeg[i]`
perhaps? or to constrain the values of BB somehow?

similarly, I notice that you program the CDF of the logistic manually, but you could have used `stats::plogis` to do the same. My guess is that the latter is more numerically stable. So for example

```r
A2 = 230
BB= 3
ddeg = 58
1 / (1+exp(-(A2-ddeg)/sqrt(BB*ddeg)))
```

would be the same as:

```r
stats::plogis(A2, location = ddeg, scale = sqrt(BB*ddeg))
```

I doubt that would actually change anything, of course! ~though it would probably get rid of the error just above.~ It probably won't help with the error just above, if it is indeed caused by `optim`'s choice of parameters and not the cdf function itself.

### The text

Really clear and readable. I was breifly confused around line 65 because the text says "this distribution" , meaning the latent logistic distribution, but equation 4 shows the CDF of this distribution, not the PDF.  You might tweak L65 to something like "The CDF in equation (4) corresponds to a logistic distribution with a mean of.." or something.

### misc questions

* You've made an historical dataset much more accessible and useful by reorganizing and documenting it. Is there a way to make the data itself more available beyond this paper?
