### 2021-04-27 pboesu/replication_candy_1991#1 @aammd
I forgot one more comment about the text!  in tables 1 and 2, I'd try to put the Method column on the left-hand side. Since the paper is about comparing these methods, the reader will want to separate the values that belong to different methods right away.
