### 2021-04-27 #46 @aammd
linked above is my review for @pboesu 's replication! To briefly summarize: I think this is excellent. Not only have you reproduced everything in the original but you've improved them in almost every aspect. The data are better organized. The math is more clearly explained. The code is modernized and well commented. The figures are identical except for some helpful colour.

Code is very clear and well-commented. I got all the errors described in the README, plus one more which had to do with parameter values tried by `optim`. I don't have much to suggest regarding the code: perhaps changing hand-coded Logistic CDFs for `stats::plogis()`.

in the text, I made one clarifying suggestion around line 65: highlight that the equation is the CDF of a latent logistic distribution, which has a mean and variance of etc.
I also suggest that in table 1 and 2, the Method column should go on the left of the table.

That's it! Thank you for the invitation, this was a pleasure.
