### 2021-04-27 #46 @juniperlsimonis
-- I'm done adding specific issues to the code repo, @pboesu! Nice job with this! --

I really like this paper and applaud you for the work on it!

I did hit some hiccups in installing and running stuff on my end, and I have some suggestions for a few ways to increase the reliability of the reproduction into the future.

- [Installation issues and improvement suggestions](https://github.com/pboesu/replication_candy_1991/issues/2)

On the stats implementation side of things, there are a few spots that I think need just a little more work, as I was able to get sufficient fits where your code failed in one spot, for example.

- [VGAM Script](https://github.com/pboesu/replication_candy_1991/issues/4)
- [Optim Sensitivity](https://github.com/pboesu/replication_candy_1991/issues/6)

I have some structure and formatting feedback that, while not critical, will improve the usability for folks in the future.

- [Code structuring and formatting suggestions](https://github.com/pboesu/replication_candy_1991/issues/3)

And then some notes on the manuscript

- [Presenting Computational Robustness of Statistical Fits](https://github.com/pboesu/replication_candy_1991/issues/5)
- [Figures](https://github.com/pboesu/replication_candy_1991/issues/7)
- [Manuscript Text](https://github.com/pboesu/replication_candy_1991/issues/8)

Again, I think this is some really great work you've done here, I just think a little bit more time spent now will make your efforts much more impactful.

This is my first review for ReScience and I really enjoyed this.

Thanks!

-Juniper Simonis (they/them/theirs)
