### 2021-04-27 pboesu/replication_candy_1991#2 @juniperlsimonis
Installation issues and improvement suggestions -- Review for ReScience
I really appreciate your attention to the installation environment, as we all know things change over time and nothing is as backwards compatible as we'd like.
I'm going to write this narratively because that's what makes sense to me, I hope you get what I'm explaining! Please let me know if you need me to clarify.

One point of note re: me is that I'm rather Ludditic with R, as I learned it in 2006 and never jumped over to R studio etc. So I honestly never deal with Rpofile files and it's not entirely clear for me what to do.  But if I open the Rprofile file in the main R gui with the plain editor, I see a familiar-to-me line:

```
source("renv/activate.R")
```

which I run and get the following

```
* Project 'C:/Users/dappe/Dropbox/reviews/replication_candy_1991' loaded. [renv 0.12.0]
* The project may be out of sync -- use `renv::status()` for more details.
Warning message:
Project requested R version '3.6.0' but '4.0.5' is currently being used
```

Those are two notes/warnings that give me pause.
I see in the README that the first one is actually not a problem --- but I didn't get that at first and had to go looking to find an explanation.

I don't see the other warning explained anywhere, just that I needed to use R 3.6.0 or later, so my thought is "ok! I need to install R 3.6.0" and so I do that and re run

```
source("renv/activate.R")
```

and it grinds to a halt during installation :

```

R version 3.6.0 (2019-04-26) -- "Planting of a Tree"
Copyright (C) 2019 The R Foundation for Statistical Computing
Platform: x86_64-w64-mingw32/x64 (64-bit)

R is free software and comes with ABSOLUTELY NO WARRANTY.
You are welcome to redistribute it under certain conditions.
Type 'license()' or 'licence()' for distribution details.

  Natural language support but running in an English locale

R is a collaborative project with many contributors.
Type 'contributors()' for more information and
'citation()' on how to cite R or R packages in publications.

Type 'demo()' for some demos, 'help()' for on-line help, or
'help.start()' for an HTML browser interface to help.
Type 'q()' to quit R.

> source("renv/activate.R")
* Downloading renv 0.12.0 from CRAN archive ... OK
* Installing renv 0.12.0 ... Done!
Successfully installed and loaded renv 0.12.0.
* Project 'C:/Users/dappe/Dropbox/reviews/replication_candy_1991' loaded. [renv 0.12.0]
* The project may be out of sync -- use `renv::status()` for more details.
> renv::restore()
The following package(s) will be updated:

# CRAN ===============================
- MASS           [7.3-51.4 -> 7.3-51.6]
- Matrix         [1.2-17 -> 1.2-18]
- lattice        [0.20-38 -> 0.20-41]
- BH             [* -> 1.72.0-3]
- R6             [* -> 2.4.1]
- RColorBrewer   [* -> 1.1-2]
- Rcpp           [* -> 1.0.5]
- VGAM           [* -> 1.1-3]
- askpass        [* -> 1.1]
- assertthat     [* -> 0.2.1]
- base64enc      [* -> 0.1-3]
- callr          [* -> 3.4.4]
- cli            [* -> 2.0.2]
- clipr          [* -> 0.7.0]
- colorspace     [* -> 1.4-1]
- cpp11          [* -> 0.2.1]
- crayon         [* -> 1.3.4]
- curl           [* -> 4.3]
- digest         [* -> 0.6.25]
- dplyr          [* -> 1.0.2]
- ellipsis       [* -> 0.3.1]
- evaluate       [* -> 0.14]
- fansi          [* -> 0.4.1]
- farver         [* -> 2.0.3]
- generics       [* -> 0.0.2]
- glue           [* -> 1.4.2]
- gtools         [* -> 3.8.2]
- highr          [* -> 0.8]
- hms            [* -> 0.5.3]
- htmltools      [* -> 0.5.0]
- httr           [* -> 1.4.2]
- jsonlite       [* -> 1.7.1]
- kableExtra     [* -> 1.2.1]
- knitr          [* -> 1.29]
- labeling       [* -> 0.3]
- latex2exp      [* -> 0.4.0]
- lifecycle      [* -> 0.2.0]
- magrittr       [* -> 1.5]
- markdown       [* -> 1.1]
- mime           [* -> 0.9]
- munsell        [* -> 0.5.0]
- numDeriv       [* -> 2016.8-1.1]
- openssl        [* -> 1.4.3]
- ordinal        [* -> 2019.12-10]
- pillar         [* -> 1.4.6]
- pkgconfig      [* -> 2.0.3]
- processx       [* -> 3.4.4]
- ps             [* -> 1.3.4]
- purrr          [* -> 0.3.4]
- readr          [* -> 1.3.1]
- rlang          [* -> 0.4.7]
- rmarkdown      [* -> 2.3]
- rstudioapi     [* -> 0.11]
- rvest          [* -> 0.3.6]
- scales         [* -> 1.1.1]
- selectr        [* -> 0.4-2]
- stringi        [* -> 1.5.3]
- stringr        [* -> 1.4.0]
- sys            [* -> 3.4]
- tibble         [* -> 3.0.3]
- tidyr          [* -> 1.1.2]
- tidyselect     [* -> 1.1.0]
- tinytex        [* -> 0.26]
- ucminf         [* -> 1.1-4]
- utf8           [* -> 1.1.4]
- vctrs          [* -> 0.3.4]
- viridisLite    [* -> 0.3.0]
- webshot        [* -> 0.5.2]
- xfun           [* -> 0.17]
- xml2           [* -> 1.3.2]
- yaml           [* -> 2.2.1]

Do you want to proceed? [y/N]: y
* Querying repositories for available binary packages ... Done!
Retrieving 'https://cran.rstudio.com/bin/windows/contrib/3.6/lattice_0.20-41.zip' ...
        OK [downloaded 1.1 Mb in 1.5 secs]
* Querying repositories for available source packages ... Done!
Retrieving 'https://cran.rstudio.com/src/contrib/Archive/MASS/MASS_7.3-51.6.tar.gz' ...
        OK [downloaded 482.6 Kb in 1.3 secs]
Retrieving 'https://mran.microsoft.com/snapshot/2020-04-24/bin/windows/contrib/3.6/Matrix_1.2-18.zip' ...
        OK [downloaded 4.3 Mb in 4.8 secs]
Retrieving 'https://cran.rstudio.com/src/contrib/Archive/Rcpp/Rcpp_1.0.5.tar.gz' ...
        OK [downloaded 2.8 Mb in 1.8 secs]
Retrieving 'https://cran.rstudio.com/src/contrib/Archive/VGAM/VGAM_1.1-3.tar.gz' ...
        OK [downloaded 2.6 Mb in 1.7 secs]
Retrieving 'https://cran.rstudio.com/bin/windows/contrib/3.6/sys_3.4.zip' ...
        OK [downloaded 58.4 Kb in 0.8 secs]
Retrieving 'https://cran.rstudio.com/src/contrib/Archive/callr/callr_3.4.4.tar.gz' ...
        OK [downloaded 74.6 Kb in 1.2 secs]
Retrieving 'https://cran.rstudio.com/src/contrib/Archive/processx/processx_3.4.4.tar.gz' ...
        OK [downloaded 131.5 Kb in 1 secs]
Retrieving 'https://cran.rstudio.com/src/contrib/Archive/ps/ps_1.3.4.tar.gz' ...
        OK [downloaded 96.7 Kb in 0.9 secs]
Retrieving 'https://cran.rstudio.com/bin/windows/contrib/3.6/glue_1.4.2.zip' ...
        OK [downloaded 151.2 Kb in 1 secs]
Retrieving 'https://cran.rstudio.com/src/contrib/Archive/cpp11/cpp11_0.2.1.tar.gz' ...
        OK [downloaded 193.5 Kb in 1.1 secs]
Retrieving 'https://cran.rstudio.com/src/contrib/Archive/dplyr/dplyr_1.0.2.tar.gz' ...
        OK [downloaded 913.7 Kb in 1.4 secs]
Retrieving 'https://cran.rstudio.com/bin/windows/contrib/3.6/ellipsis_0.3.1.zip' ...
        OK [downloaded 45.1 Kb in 0.8 secs]
Retrieving 'https://cran.rstudio.com/src/contrib/Archive/rlang/rlang_0.4.7.tar.gz' ...
        OK [downloaded 827.4 Kb in 1.5 secs]
Retrieving 'https://cran.rstudio.com/src/contrib/Archive/tibble/tibble_3.0.3.tar.gz' ...
        OK [downloaded 248.6 Kb in 1 secs]
Retrieving 'https://cran.rstudio.com/src/contrib/Archive/pillar/pillar_1.4.6.tar.gz' ...
        OK [downloaded 111.2 Kb in 0.9 secs]
Retrieving 'https://cran.rstudio.com/src/contrib/Archive/vctrs/vctrs_0.3.4.tar.gz' ...
        OK [downloaded 971.8 Kb in 1.4 secs]
Retrieving 'https://cran.rstudio.com/bin/windows/contrib/3.6/tidyselect_1.1.0.zip' ...
        OK [downloaded 198.5 Kb in 1.1 secs]
Retrieving 'https://cran.rstudio.com/bin/windows/contrib/3.6/purrr_0.3.4.zip' ...
        OK [downloaded 420.1 Kb in 1.2 secs]
Retrieving 'https://cran.rstudio.com/src/contrib/Archive/htmltools/htmltools_0.5.0.tar.gz' ...
        OK [downloaded 64.5 Kb in 0.8 secs]
Retrieving 'https://cran.rstudio.com/bin/windows/contrib/3.6/httr_1.4.2.zip' ...
        OK [downloaded 506.8 Kb in 1.3 secs]
Retrieving 'https://cran.rstudio.com/src/contrib/Archive/jsonlite/jsonlite_1.7.1.tar.gz' ...
        OK [downloaded 1 Mb in 1.4 secs]
Retrieving 'https://cran.rstudio.com/bin/windows/contrib/3.6/openssl_1.4.3.zip' ...
        OK [downloaded 3.8 Mb in 2 secs]
Retrieving 'https://cran.rstudio.com/src/contrib/Archive/kableExtra/kableExtra_1.2.1.tar.gz' ...
        OK [downloaded 1.6 Mb in 1.6 secs]
Retrieving 'https://cran.rstudio.com/src/contrib/Archive/knitr/knitr_1.29.tar.gz' ...
        OK [downloaded 864.9 Kb in 1.5 secs]
Retrieving 'https://cran.rstudio.com/src/contrib/Archive/xfun/xfun_0.17.tar.gz' ...
        OK [downloaded 75.4 Kb in 1 secs]
Retrieving 'https://cran.rstudio.com/bin/windows/contrib/3.6/stringi_1.5.3.zip' ...
        OK [downloaded 14.6 Mb in 3.4 secs]
Retrieving 'https://cran.rstudio.com/bin/windows/contrib/3.6/xml2_1.3.2.zip' ...
        OK [downloaded 3 Mb in 1.8 secs]
Retrieving 'https://cran.rstudio.com/src/contrib/Archive/rvest/rvest_0.3.6.tar.gz' ...
        OK [downloaded 2.7 Mb in 1.7 secs]
Retrieving 'https://cran.rstudio.com/bin/windows/contrib/3.6/selectr_0.4-2.zip' ...
        OK [downloaded 479.4 Kb in 1.4 secs]
Retrieving 'https://cran.rstudio.com/src/contrib/Archive/rmarkdown/rmarkdown_2.3.tar.gz' ...
        OK [downloaded 3 Mb in 1.8 secs]
Retrieving 'https://cran.rstudio.com/src/contrib/Archive/tinytex/tinytex_0.26.tar.gz' ...
        OK [downloaded 26.4 Kb in 0.7 secs]
Retrieving 'https://cran.rstudio.com/bin/windows/contrib/3.6/scales_1.1.1.zip' ...
        OK [downloaded 548.4 Kb in 1.3 secs]
Retrieving 'https://mran.microsoft.com/snapshot/2020-04-24/bin/windows/contrib/3.6/latex2exp_0.4.0.zip' ...
        OK [downloaded 313.6 Kb in 1.8 secs]
Retrieving 'https://cran.rstudio.com/bin/windows/contrib/3.6/numDeriv_2016.8-1.1.zip' ...
        OK [downloaded 113.7 Kb in 0.9 secs]
Retrieving 'https://cran.rstudio.com/bin/windows/contrib/3.6/ordinal_2019.12-10.zip' ...
        OK [downloaded 1.2 Mb in 1.6 secs]
Retrieving 'https://cran.rstudio.com/bin/windows/contrib/3.6/ucminf_1.1-4.zip' ...
        OK [downloaded 187 Kb in 1 secs]
Retrieving 'https://cran.rstudio.com/src/contrib/Archive/tidyr/tidyr_1.1.2.tar.gz' ...
        OK [downloaded 860.9 Kb in 1.4 secs]
Installing lattice [0.20-41] ...
        OK [installed binary]
Installing MASS [7.3-51.6] ...
        FAILED
Error installing package 'MASS':
================================

* installing *source* package 'MASS' ...
** package 'MASS' successfully unpacked and MD5 sums checked
** using staged installation
** libs
Warning in system(paste(cmd, "shlib-clean")) : 'make' not found
Warning in system(cmd) : 'make' not found
ERROR: compilation failed for package 'MASS'
* removing 'C:/Users/dappe/Dropbox/reviews/REPLIC~1/renv/staging/1/MASS'
Error: install of package 'MASS' failed
Traceback (most recent calls last):
14: renv::restore()
13: renv_restore_run_actions(project, diff, current, lockfile)
12: renv_install(records, library)
11: renv_install_staged(records, library)
10: renv_install_default(records, library)
 9: handler(package, renv_install_impl(record))
 8: renv_install_impl(record)
 7: withCallingHandlers(renv_install_package_local(record), error = function(e) {
        vwritef("\tFAILED")
        writef(e$output)
    })
 6: renv_install_package_local(record)
 5: renv_install_package_local_impl(package, path, library)
 4: r_cmd_install(package, path, library)
 3: r_exec(package, args, "install")
 2: r_exec_error(package, output, label)
 1: stop(error)
>
```

So that was a bummer! But I was thinking I'd still try to run the code in R4.0.5, given that it worked but just gave me a warning. And it works!

Suggestions:
- rearrange top-level dependency loading code to align with other code
  - there's no reason that the dependency loading code (done during one source call) needs to be separated from the other scripts that are also sourced
  - from a code understanding standpoint, it's just ideal to have things follow a similar mechanism
- provide complete session / run environment information, as that will allow someone to recreate the environment you used in the case of any installation breaks
  - `sessionInfo()` will provide this
  - the ultimate implementation would be to create an image of the runtime environment (e.g. via Docker), but that is not necessary for this project, as long as someone can recreate the environment
- issues with dependency hell (almost certainly what i hit above there) are going to be more likely the more you rely on non-base packages to manage installations
  - if you can restructure the installation to not use **renv**, that will improve reproducibility
- mute any warnings or notes that are expected and not meaningful to the user
  - This is not a "mute all warnings" or whatnot, but specifically those that you would be telling the user to ignore, it makes the user experience better when they just don't see them in the first place.

Here is the sessionInfo() for my session that worked fine

```

> sessionInfo()
R version 4.0.5 (2021-03-31)
Platform: x86_64-w64-mingw32/x64 (64-bit)
Running under: Windows 10 x64 (build 19041)

Matrix products: default

locale:
[1] LC_COLLATE=English_United States.1252  LC_CTYPE=English_United States.1252
[3] LC_MONETARY=English_United States.1252 LC_NUMERIC=C
[5] LC_TIME=English_United States.1252

attached base packages:
[1] splines   stats4    stats     graphics  grDevices utils     datasets  methods   base

other attached packages:
[1] kableExtra_1.3.4   ordinal_2019.12-10 VGAM_1.1-4         dplyr_1.0.3

 [1] latex2exp_0.5.0     pillar_1.4.7        compiler_4.0.5      tools_4.0.5         digest_0.6.27
 [6] viridisLite_0.3.0   evaluate_0.14       lifecycle_1.0.0     tibble_3.0.5        lattice_0.20-41
[11] ucminf_1.1-4        pkgconfig_2.0.3     rlang_0.4.10        Matrix_1.3-2        cli_2.4.0
[16] rstudioapi_0.13     xfun_0.21           stringr_1.4.0       httr_1.4.2          knitr_1.31
[21] xml2_1.3.2          systemfonts_1.0.1   generics_0.1.0      vctrs_0.3.6         gtools_3.8.2
[26] hms_0.5.3           webshot_0.5.2       grid_4.0.5          tidyselect_1.1.0    svglite_2.0.0
[31] glue_1.4.2          R6_2.5.0            rmarkdown_2.7       tidyr_1.1.2         readr_1.4.0
[36] purrr_0.3.4         magrittr_2.0.1      scales_1.1.1        ps_1.5.0            ellipsis_0.3.1
[41] htmltools_0.5.1.1   MASS_7.3-53.1       rvest_1.0.0         colorspace_2.0-0    numDeriv_2016.8-1.1
[46] stringi_1.4.3       munsell_0.5.0       crayon_1.3.4
```
