### 2021-04-27 pboesu/replication_candy_1991#3 @juniperlsimonis
Code structuring and formatting suggestions -- Review for ReScience
A lot of what I'm writing here is about personal preferences and such, so keep that in mind, but I'm trying to think of this from the sake that a wide range of users might be interested in reproducing the work, so having stuff be generally accessible is what I'm trying to target.

I really like that stuff is pulled out into different scripts to help organize the codebase
I think there are some ways to refine this just a bit more.
In particular, I would really isolate the code that a user is expected to run at the console vs. what the console is sourcing.
For example, you tell users in the README to just source the scripts, but that's actually not how those scripts are really set up.
There's a lot of stuff commented in the code that's important to know, and some of the stuff runs and some of it fails. Indeed, when I follow your README I get the following

```
> source('code/01_reformat_budworm_data.R')
> source('code/02_fit_candy_1991_models.R')
Parsed with column specification:
cols(
  ddeg = col_double(),
  total = col_double(),
  stage = col_double(),
  count = col_double()
)
initial  value 94150.793409
iter  10 value 187.831975
iter  20 value 92.623925
final  value 92.584687
converged
initial  value 106684.130833
iter  10 value 234.717054
iter  20 value 97.619139
iter  30 value 96.373309
iter  30 value 96.373309
final  value 96.323276
converged
initial  value 1375.973037
iter  10 value 81.150500
iter  20 value 81.149096
final  value 81.149089
converged
There were 25 warnings (use warnings() to see them)

> source('code/03_fit_dennis_et_al_model.R')

Attaching package: ‘kableExtra’

The following object is masked from ‘package:dplyr’:

    group_rows

Parsed with column specification:
cols(
  ddeg = col_double(),
  total = col_double(),
  stage = col_double(),
  count = col_double()
)
initial  value 500.857239
iter  10 value 408.602852
iter  20 value 407.440856
iter  30 value 407.380269
final  value 407.380044
converged
There were 50 or more warnings (use warnings() to see the first 50)

Loading required package: stats4
Loading required package: splines
Parsed with column specification:
cols(
  ddeg = col_double(),
  total = col_double(),
  stage1 = col_double(),
  stage2 = col_double(),
  stage3 = col_double(),
  stage4 = col_double(),
  stage5 = col_double(),
  stage6 = col_double(),
  stage7 = col_double(),
  ddeg_cent = col_double()
)
Error in checkwz(wz, M, trace = trace, wzepsilon = control$wzepsilon) :
  NAs found in the working weights variable 'wz'
In addition: Warning message:
In Deviance.categorical.data.vgam(mu = mu, y = y, w = w, residuals = residuals,  :
  fitted values close to 0 or 1
VGLM    linear loop  1 :  deviance = 385.77698
Error in checkwz(wz, M = M, trace = trace, wzepsilon = control$wzepsilon) :
  NAs found in the working weights variable 'wz'
In addition: Warning messages:
1: In checkwz(wz, M = M, trace = trace, wzepsilon = control$wzepsilon) :
  5 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
2: In checkwz(wz, M = M, trace = trace, wzepsilon = control$wzepsilon) :
  7 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
3: In checkwz(wz, M = M, trace = trace, wzepsilon = control$wzepsilon) :
  10 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
4: In checkwz(wz, M = M, trace = trace, wzepsilon = control$wzepsilon) :
  12 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
5: In checkwz(wz, M = M, trace = trace, wzepsilon = control$wzepsilon) :
  12 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
6: In checkwz(wz, M = M, trace = trace, wzepsilon = control$wzepsilon) :
  12 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
7: In checkwz(wz, M = M, trace = trace, wzepsilon = control$wzepsilon) :
  12 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
8: In checkwz(wz, M, trace = trace, wzepsilon = control$wzepsilon) :
  46 diagonal elements of the working weights variable 'wz' have been replaced by 0.5
9: In tfun(mu = mu, y = y, w = w, res = FALSE, eta = eta, extra = extra) :
  fitted values close to 0 or 1

> source('code/05_fit_ordinal_models.R')

Attaching package: ‘ordinal’

The following objects are masked from ‘package:VGAM’:

    dgumbel, dlgamma, pgumbel, plgamma, qgumbel, rgumbel, wine

The following object is masked from ‘package:dplyr’:

    slice

Parsed with column specification:
cols(
  ddeg = col_double(),
  total = col_double(),
  stage = col_double(),
  ddeg_cent = col_double()
)

> source('code/06_collate_parameter_tables.R')
```

Which is to say that all by 04 ran without error, but 02, and 03 throw warnings and 02, 03, and 05 print messages, but not related to what a user would want

But 04 did not run because of the errors, which means that a user necessarily has to do more than what you say in the README and go find out why.

Something to improve the usability here would be to again quiet the extensive messaging that readr throws out. The less the user sees at the console of things they dont need to see, the better.

I'm all for having relevant info presented to the user, but I think a lot of what the user sees isn't meaningful.
I'm not a big fan of telling the user to just not worry about warnings that come to the console, so as I mentioned previously, I'd suggest either addressing those issues so the warnings go away or silence them specifically.

So yeah, the idea would be, rather than rely on the README for the user to follow for everything including code running, create a simple top-level script that
1. loads the environment
2. sources the internal function
3. walks through the analyses

This would remove the special Rprofile nature of the installation and align everything a user would want to read in the code for a first run through
This gets to another part of the concern I had, which is that it forces users to bounce between the source (which gives very little info and only via file names) in the README to the actual files, which have more details, but perhaps are a bit overwhelming

To address some of that, I'd suggest:
1. Any comment lines should have a space after the # character to help with reading
  - In general, take advantage of white space as much as you can to help new users read your code
2.  Split function creation from function usage
  - Presently, the functions and the code using them are together in the .R files, which makes it challenging to navigate.
  - A simplifying format/structure would be to condense the functions into their own script (perhaps sectioned off by topic or if desired with a script for each topic) and then source those at the beginning of the top-level script (see 2. in previous list)
    - This allows the user to get the gist of what you're doing, and then dive into the function details if they want
    - The top level script that walks through the analyses (3. in previous list) can have a little table of contents and explanation in comments, which is why it's useful to have it separated from the README

This will circumvent the need for repeated calls to loading libraries in each sourced script

This also addresses an issue regarding the things that *are* printed to the screen within the scripts, but don't get echoed out when you source them, like

```
> logit_cm_candy_nll$par
[1]  5.53478517  9.48212482 12.35402857 15.85841654 21.46339190 27.51646597 -0.04613333
```

If that's something the user should be seeing, make sure they're actually seeing it!

What is the justification for saving things as .RDS files? If you don't care that much one way or another, I would suggest saving this kind of content in text type files, be they actual text, json/yml/whatever, it's just much more reliable and transferable (i shouldn't need R to open them).
