### 2021-04-27 pboesu/replication_candy_1991#4 @juniperlsimonis
VGAM Script -- Review for ReScience
Having gotten the error (#3) in the VGAM script, I dug into it a bit more to see the situation

Here's what my running of the script returns:

```
> library(VGAM)
> budworm_table <- readr::read_csv('data/budworm_table.csv')
Parsed with column specification:
cols(
  ddeg = col_double(),
  total = col_double(),
  stage1 = col_double(),
  stage2 = col_double(),
  stage3 = col_double(),
  stage4 = col_double(),
  stage5 = col_double(),
  stage6 = col_double(),
  stage7 = col_double(),
  ddeg_cent = col_double()
)
>
> #fit cumulative model
> budworm_cumulative_cloglog_vglm <- try(vglm(cbind(stage1, stage2, stage3, stage4, stage5, stage6, stage7) ~ ddeg,
+                                         cumulative(reverse = FALSE, link = clogloglink(bvalue = 1e-8), parallel = TRUE), coefstart = c(-3,-6,-8,-10,-14,-17,0.01), data = budworm_table, trace=T))#fails
Error in checkwz(wz, M, trace = trace, wzepsilon = control$wzepsilon) :
  NAs found in the working weights variable 'wz'
In addition: Warning message:
In Deviance.categorical.data.vgam(mu = mu, y = y, w = w, residuals = residuals,  :
  fitted values close to 0 or 1
>
>
> budworm_cumulative_logit_vglm <- vglm(cbind(stage1, stage2, stage3, stage4, stage5, stage6, stage7) ~ ddeg,
+                                       cumulative(reverse = FALSE, link = "logitlink", parallel = TRUE), data = budworm_table) #reproduces parameter estimates in table 2 of candy
>
> #save parameter estimates
> candy_vglm_cm_estimates <- as.data.frame(rbind(unname(coef(budworm_cumulative_logit_vglm)), rep(NA,7)))
> candy_vglm_cm_estimates$model <- "cumulative"
> candy_vglm_cm_estimates$link <- c("logit","cloglog")
> candy_vglm_cm_estimates$fit <- c("R \\verb+vglm+")
> candy_vglm_cm_estimates$eqn <- c(NA,NA)
> saveRDS(candy_vglm_cm_estimates, "outputs/candy_vglm_cm_estimates.RDS")
> #create predictions to compare with candy original fit
> #pred_data <- expand.grid(ddeg = 0:800, stage = 1:7, total = 100)
> #pred_data$alpha_index = pred_data$stage + 1
>
> #fit sequential model
> budworm_sratio_logit_vglm <- vglm(cbind(stage1, stage2, stage3, stage4, stage5, stage6, stage7) ~ ddeg,
+                                   sratio(link = "logitlink", parallel = FALSE), data = budworm_table)
Warning messages:
1: In checkwz(wz, M = M, trace = trace, wzepsilon = control$wzepsilon) :
  5 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
2: In checkwz(wz, M = M, trace = trace, wzepsilon = control$wzepsilon) :
  7 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
3: In checkwz(wz, M = M, trace = trace, wzepsilon = control$wzepsilon) :
  10 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
4: In checkwz(wz, M = M, trace = trace, wzepsilon = control$wzepsilon) :
  12 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
5: In checkwz(wz, M = M, trace = trace, wzepsilon = control$wzepsilon) :
  12 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
6: In checkwz(wz, M = M, trace = trace, wzepsilon = control$wzepsilon) :
  12 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
7: In checkwz(wz, M = M, trace = trace, wzepsilon = control$wzepsilon) :
  12 diagonal elements of the working weights variable 'wz' have been replaced by 1.819e-12
>
> #t(coef(budworm_sratio_logit_vglm, matrix = TRUE))
>
> budworm_sratio_cloglog_vglm <- try(vglm(cbind(stage1, stage2, stage3, stage4, stage5, stage6, stage7) ~ ddeg_cent,
+                                         sratio(link = "clogloglink", parallel = FALSE), data = budworm_table, trace=TRUE, control = vglm.control(checkwz = TRUE, wzepsilon = 0.5)))#sratio model with clogloglink fails because of some numerical issue - but check the expected binomial sample sizes. Candy writes "Note that the n_ij for which the binomial sample size is N^*_ij is zero must be expluded from the fit by by giving them a prior weight of zero. VGAM requires positive weights but suggests using weights "such as 1e-8" to effectively exclude observations"
VGLM    linear loop  1 :  deviance = 385.77698
Error in checkwz(wz, M = M, trace = trace, wzepsilon = control$wzepsilon) :
  NAs found in the working weights variable 'wz'
In addition: Warning messages:
1: In checkwz(wz, M, trace = trace, wzepsilon = control$wzepsilon) :
  46 diagonal elements of the working weights variable 'wz' have been replaced by 0.5
2: In tfun(mu = mu, y = y, w = w, res = FALSE, eta = eta, extra = extra) :
  fitted values close to 0 or 1
>
> #save parameter estimates
> candy_vglm_sm_estimates <- as.data.frame(rbind(unname(coef(budworm_sratio_logit_vglm)[1:6]),unname(coef(budworm_sratio_logit_vglm)[7:12]),rep(NA,6),rep(NA,6)))
> candy_vglm_sm_estimates$model <- "sequential"
> candy_vglm_sm_estimates$link <- c("logit","logit","cloglog","cloglog")
> candy_vglm_sm_estimates$fit <- c("R \\verb+vglm+")
> candy_vglm_sm_estimates$par <- c("$\\beta_{0j}$","$\\beta_{1j}$","$\\beta_{0j}$","$\\beta_{1j}$")
> saveRDS(candy_vglm_sm_estimates, "outputs/candy_vglm_sm_estimates.RDS")
>

```

It looks like you have a `try` wrapped about at least one of the one that fails, but perhaps the arguments need to be edited?

Further playing around with the model function leads me to a concern regarding the success/failure attributed to these models:

```
> budworm_cumulative_cloglog_vglm <- try(vglm(cbind(stage1, stage2, stage3, stage4, stage5, stage6, stage7) ~ ddeg,
+                                         cumulative(reverse = FALSE, link = clogloglink(bvalue = 1e-8), parallel = TRUE), coefstart = c(-3,-6,-8,-10,-14,-17,0.01), data = budworm_table, trace=T))#fails
Error in checkwz(wz, M, trace = trace, wzepsilon = control$wzepsilon) :
  NAs found in the working weights variable 'wz'
In addition: Warning message:
In Deviance.categorical.data.vgam(mu = mu, y = y, w = w, residuals = residuals,  :
  fitted values close to 0 or 1
>
```

vs

```
> budworm_cumulative_cloglog_vglm <- try(vglm(cbind(stage1, stage2, stage3, stage4, stage5, stage6, stage7) ~ ddeg,
+                                         cumulative(reverse = FALSE, link = clogloglink(bvalue = 1e-8), parallel = TRUE),  data = budworm_table, trace=T))#fails
VGLM    linear loop  1 :  deviance = 1216.7404
VGLM    linear loop  2 :  deviance = 569.75062
Taking a modified step...
VGLM    linear loop  2 :  deviance = 1086.8542
VGLM    linear loop  3 :  deviance = 512.86298
Taking a modified step.....
VGLM    linear loop  3 :  deviance = 1057.2014
VGLM    linear loop  4 :  deviance = 499.18205
Taking a modified step.......
VGLM    linear loop  4 :  deviance = 1049.9362
VGLM    linear loop  5 :  deviance = 495.79338
Taking a modified step.........
VGLM    linear loop  5 :  deviance = 1048.1288
VGLM    linear loop  6 :  deviance = 494.94817
Taking a modified step..........
VGLM    linear loop  6 :  deviance = 1047.2265
VGLM    linear loop  7 :  deviance = 494.5259
Taking a modified step...........
VGLM    linear loop  7 :  deviance = 1046.7756
VGLM    linear loop  8 :  deviance = 494.31486
Taking a modified step............
VGLM    linear loop  8 :  deviance = 1046.5503
VGLM    linear loop  9 :  deviance = 494.20936
Taking a modified step..............
VGLM    linear loop  9 :  deviance = 1046.494
VGLM    linear loop  10 :  deviance = 494.18298
Taking a modified step.................
VGLM    linear loop  10 :  deviance = 1046.4869
VGLM    linear loop  11 :  deviance = 494.17969
Taking a modified step..................
VGLM    linear loop  11 :  deviance = 1046.4834
VGLM    linear loop  12 :  deviance = 494.17804
Taking a modified step....................
There were 50 or more warnings (use warnings() to see the first 50)
```

All I did was remove the starting values you had used, which I imagine sets them to some other default or something, I didn't follow up on the specific results, the point here mainly being that I believe you might actually be able to fit the models that you say are failing if you do a little more work around the starting values and potentially other arguments.

I'm not entirely sure about the second message because the formulation of VGAM stuff is new to me, so I didn't get too far in digging, but it appears similar, so I would imagine there may be starting value issues here, as well.

Unless I'm reading it wrong, the starting values weren't discussed in the OG MSes and you're using a new method for implementation here (VGAM) so, I think the onus is on you to do more exploration before declaring these failures.
