### 2021-04-27 pboesu/replication_candy_1991#5 @juniperlsimonis
Presenting Computational Robustness of Statistical Fits -- Review for ReScience
I'm a bit concerned about the robustness (or lack) of some of the statistical fits, which does not get mentioned in the manuscript (or if it does, needs to be more discussed)

For example

```
> cloglog_sm_candy_glm <- glm(cbind(count, n_star-count) ~ stage_factor:ddeg + stage_factor-1, family = binomial(link="cloglog"), data = budworm_counts, subset = stage < 7)
Warning message:
glm.fit: fitted probabilities numerically 0 or 1 occurred
> summary(cloglog_sm_candy_glm )

Call:
glm(formula = cbind(count, n_star - count) ~ stage_factor:ddeg +
    stage_factor - 1, family = binomial(link = "cloglog"), data = budworm_counts,
    subset = stage < 7)

Deviance Residuals:
     Min        1Q    Median        3Q       Max
-1.80130  -0.05384   0.00000   0.00000   1.81496

Coefficients:
                    Estimate Std. Error z value Pr(>|z|)
stage_factor1       7.347006   1.419074   5.177 2.25e-07 ***
stage_factor2       8.537078   1.197913   7.127 1.03e-12 ***
stage_factor3       9.124225   1.024015   8.910  < 2e-16 ***
stage_factor4       8.442036   0.937439   9.005  < 2e-16 ***
stage_factor5      10.086804   0.962310  10.482  < 2e-16 ***
stage_factor6      16.297979   2.040615   7.987 1.38e-15 ***
stage_factor1:ddeg -0.065109   0.012423  -5.241 1.60e-07 ***
stage_factor2:ddeg -0.044296   0.005659  -7.828 4.97e-15 ***
stage_factor3:ddeg -0.036766   0.003969  -9.263  < 2e-16 ***
stage_factor4:ddeg -0.026290   0.002749  -9.565  < 2e-16 ***
stage_factor5:ddeg -0.022841   0.002213 -10.323  < 2e-16 ***
stage_factor6:ddeg -0.028543   0.003596  -7.938 2.06e-15 ***
---
Signif. codes:  0 ‘***’ 0.001 ‘**’ 0.01 ‘*’ 0.05 ‘.’ 0.1 ‘ ’ 1

(Dispersion parameter for binomial family taken to be 1)

    Null deviance: 4265.646  on 50  degrees of freedom
Residual deviance:   23.888  on 38  degrees of freedom
AIC: 107.38

Number of Fisher Scoring iterations: 10

>

```

It's totally understandable to get these kinds of warnings with ecological data, I'm not saying it's anything novel or something to be super concerned about, BUT, I do think it needs to be conveyed to the reader

As it stands now, readers/users are basically told to ignore the warnings, but these are actually important warnings that should be considered (even if not "acted on").
