### 2021-04-27 pboesu/replication_candy_1991#6 @juniperlsimonis
Optim Sensitivity -- Review for ReScience
I appreciate the acknowledgement of the importance of starting values for the results in optim, but it would be really useful to have
1. information the range of parameters used
2. some of those results displayed
3.  code showing the user how you did that exploration.

Is it possible to include that?
