### 2021-04-27 pboesu/replication_candy_1991#7 @juniperlsimonis
Figures -- Review for ReScience
Nice job with the figures!

A few quick notes:

you're saving them out as Fig 2 and Fig 3 but using them as Fig 1 and Fig 2 --- suggest aligning

note on them or in the legend what specific replication you're matching, as you show multiple model fits for these.
