### 2021-04-27 pboesu/replication_candy_1991#8 @juniperlsimonis
Manuscript Text -- Review for ReScience
I think the text of the MS flows nicely and is generally easy to read.

A few quick additional notes/questions

- Include the specific optimization routines being used in the model fits. Optim has multiple options. If you used the default, make sure to include that.
- Similarly, with all of the functions, tell us more of the arguments you used/any decisions around that. Especially in the context of the minor differences you found, what did you explore and how?
- Contextualize the degree of parameter difference you mention on line 129: relatively speaking, what does this discrepancy mean?
