### 2021-05-03 #46 @rougier
@aammd Woudl you mind posting your review in this thread? Once published, we link the paper to the review and if for some reason the repository where you posted your review become unavailable, your review will be lost. Here (ReScience/submission) we can guarantee future access (somehow).
