### 2021-05-15 pboesu/replication_candy_1991#1 @pboesu
Checklist for me to track these comments

- [x]  `optim` constraints - addressed in https://github.com/pboesu/replication_candy_1991/commit/f5e4646bd01b4170768666926df4472997787447
- [x]  `stats::plogis` vs. literal translation of original code - adressed in https://github.com/pboesu/replication_candy_1991/commit/f5e4646bd01b4170768666926df4472997787447
- [x]  text: L65 -  addressed in https://github.com/pboesu/replication_candy_1991/commit/c189e55e4d50dd89e6b73c17787399f52c447344
- [x]  text: table 1,2 column order - adressed in https://github.com/pboesu/replication_candy_1991/commit/a199c1c0a92eab2f46e6d51968459d0eccb690ad
- [ ]  data set availability
