### 2021-05-16 97c67d6
switch optim to L-BFGS-B and use positive  constraint on parameters
