### 2021-05-16 f5e4646
add plogis implementation of Dennis model
