+++ code/03_fit_dennis_et_al_model.R
@@ -2,7 +2,7 @@ library(dplyr)
 library(kableExtra)
 budworm_counts <- readr::read_csv('data/budworm_counts.csv')
 
-#optimize likelihood directly rather than doing the IRLS approach
+#optimize likelihood directly rather than doing the IRLS approach. This is a quite literal translation of the original SAS code.
 nll_cm_dennis <- function(par, count, total, stage, ddeg){
   A1 <- par[1]
   A2 <- par[2]
@@ -26,6 +26,32 @@ nll_cm_dennis <- function(par, count, total, stage, ddeg){
   return(nll)
 }
 logit_cm_dennis_nll <- optim(par = c(A1 = 150, A2 = 230, A3 = 280, A4 = 330, A5 = 440, A6 = 580, BB = 3), nll_cm_dennis, count = budworm_counts$count, total = budworm_counts$total, stage = budworm_counts$stage, ddeg = budworm_counts$ddeg, hessian = TRUE, control = list(trace=1), method = 'L-BFGS-B', lower = rep(0,7))
+
+#This is an alternative implementation that uses the R function stats::plogis 
+nll_cm_dennis_plogis <- function(par, count, total, stage, ddeg){
+  A1 <- par[1]
+  A2 <- par[2]
+  A3 <- par[3]
+  A4 <- par[4]
+  A5 <- par[5]
+  A6 <- par[6]
+  BB <- par[7]
+  pred <- weight <- rep(NA, length(total))
+  for (i in seq_along(total)){
+    pred[i] <- (stage[i]==1)*stats::plogis(A1, location = ddeg[i], scale = sqrt(BB*ddeg[i]))+
+      (stage[i]==2)*(stats::plogis(A2, location = ddeg[i], scale = sqrt(BB*ddeg[i]))-stats::plogis(A1, location = ddeg[i], scale = sqrt(BB*ddeg[i])))+
+      (stage[i]==3)*(stats::plogis(A3, location = ddeg[i], scale = sqrt(BB*ddeg[i]))-stats::plogis(A2, location = ddeg[i], scale = sqrt(BB*ddeg[i])))+
+      (stage[i]==4)*(stats::plogis(A4, location = ddeg[i], scale = sqrt(BB*ddeg[i]))-stats::plogis(A3, location = ddeg[i], scale = sqrt(BB*ddeg[i])))+
+      (stage[i]==5)*(stats::plogis(A5, location = ddeg[i], scale = sqrt(BB*ddeg[i]))-stats::plogis(A4, location = ddeg[i], scale = sqrt(BB*ddeg[i])))+
+      (stage[i]==6)*(stats::plogis(A6, location = ddeg[i], scale = sqrt(BB*ddeg[i]))-stats::plogis(A5, location = ddeg[i], scale = sqrt(BB*ddeg[i])))+
+      (stage[i]==7)*(stats::plogis(A6, location = ddeg[i], scale = sqrt(BB*ddeg[i]), lower.tail = FALSE))
+  }
+  pred <- pmax(pred, 1e-8)
+  nll <- -1*sum(count*log(pred))
+  return(nll)
+}
+
+logit_cm_dennis_nll_plogis <- optim(par = c(A1 = 150, A2 = 230, A3 = 280, A4 = 330, A5 = 440, A6 = 580, BB = 3), nll_cm_dennis_plogis, count = budworm_counts$count, total = budworm_counts$total, stage = budworm_counts$stage, ddeg = budworm_counts$ddeg, hessian = TRUE, control = list(trace=1, maxit = 500), method = 'L-BFGS-B', lower = rep(1e-12,7))
 #logit_cm_dennis_nll <- add_se_vcov_nll_hessian(logit_cm_dennis_nll)
 
[…]
