### 2021-05-16 e99ba55
add optim method details to ms
