+++ article/bibliography.bib
@@ -303,3 +303,14 @@ doi = {10.1086/705241}
     note = {R package version 3.8.2},
     url = {https://CRAN.R-project.org/package=gtools},
   }
+  
+@article{byrd1995limited,
+  title={A limited memory algorithm for bound constrained optimization},
+  author={Byrd, Richard H and Lu, Peihuang and Nocedal, Jorge and Zhu, Ciyou},
+  journal={SIAM Journal on scientific computing},
+  volume={16},
+  number={5},
+  pages={1190--1208},
+  year={1995},
+  publisher={SIAM}
+}
