### 2021-05-16 6108fdf
mention both optim approaches for Dennis model in results
