### 2021-05-16 pboesu/replication_candy_1991#8 @pboesu
Revisions checklist

- [x] add details of `optim` method  - adressed in e99ba55
- [x] broader context for other function arguments - added statements that default settings were used in all fittting functions other than `optim`
- [x] quantify parameter differences between original and replication - adressed in 644ab1438ea7413decc6dbdddd3882d3c5e51686
