### 2021-05-16 92dec06
clarify figure naming and model fits underlying figures
