+++ article/content.tex
@@ -185,10 +185,10 @@ Figure~\ref{fig:fig1} and~\ref{fig:fig2}, respectively,  show that despite minor
  
 \begin{figure}[p]
   \centering
-  \includegraphics[width=\textwidth]{../figures/dennis_fig2.pdf}
+  \includegraphics[width=\textwidth]{../figures/fig1_dennis_fig2.pdf}
   \caption{Logistic PDF of the cumulative model with proportional variance (Equation~\ref{eq:dennis_cm}) plotted for fixed values of~$t$. 
   Area under the PDF between $a_{j-1}$ and $a_j$ gives the expected proportion of insects in stage $j$ at time $t$. 
-  The graph is based on the estimates in Table 1 of \citep{kemp1986stochastic} (black lines) and the estimates from the replication (red lines). 
+  The graph is based on the estimates in Table 1 of \citep{kemp1986stochastic} (black lines) and the estimates from the replication using the R \verb+optim+ implementation (red lines). 
   This figure replicates Figure 2 in \citep{dennis1986stochastic}.}
   \label{fig:fig1}
 \end{figure}
@@ -196,9 +196,9 @@ Figure~\ref{fig:fig1} and~\ref{fig:fig2}, respectively,  show that despite minor
 
 \begin{figure}[p]
   \centering
-  \includegraphics[width=\textwidth]{../figures/dennis_fig3.pdf}
+  \includegraphics[width=\textwidth]{../figures/fig2_dennis_fig3.pdf}
   \caption{Proportion of insects expected in stages 1-7 under the cumulative logit model with proportional variance (Equation~\ref{eq:dennis_cm}) plotted as functions of time $t$.
-   Values of $a_j$ and $b^2$ used in the graph are the estimates given in Table 1 of \citep{kemp1986stochastic} (black lines) and the estimates from the replication (red lines). 
+   Values of $a_j$ and $b^2$ used in the graph are the estimates given in Table 1 of \citep{kemp1986stochastic} (black lines) and the estimates from the replication using the R \verb+optim+ implementation (red lines). 
    This figure replicates Figure 3 in \citep{dennis1986stochastic}.}
   \label{fig:fig2}
 \end{figure} 
