### 2021-05-16 pboesu/replication_candy_1991#7 @pboesu
https://github.com/pboesu/replication_candy_1991/commit/92dec06d96024669aa18e54578b3848b749da325 - clarifies the figure naming in the code and file names as well as the replication implementation underlying the figures. These are figures 1 and 2 in the replication manuscript, but figures 2 and 3, respectively in the source manuscript.
