### 2021-05-17 pboesu/replication_candy_1991#4 @pboesu
Revisions checklist

- [x] reproduce `VGAM` behaviour w/o starting values
- [x]  email with error description and reprex sent to `VGAM` maintainer on 2021-05-17 - reprex at https://gist.github.com/pboesu/bfa6626b0b56e887d84f43bc6f08590b
- [x] expand write up to state that this has been reported as a bug 94e71ff5a84972b32fd5ce2f464d9f9f0f7e7a53 , d227d2eef0a337a2b02fb054e451cbbeec1c9e6e which is yet to be fully resolved
- [x] added additional context about the dataset to the manuscript, and how its sparsity affects the numerical  stability of model estimation, e.g.  7338d290d433656d1ec2af7c355ee41d608613da 94e71ff5a84972b32fd5ce2f464d9f9f0f7e7a53
