### 2021-05-18 6e0662d
constrain predictions to be positive to avoid numerical underflow / infinite values in the poisson likelihood
