+++ article/article.tex
@@ -13,5 +13,5 @@
 \renewcommand\emph[1]{{\bfseries #1}}
 \setlength\bibitemsep{0pt}
 \printbibliography
-
+\input{appendix.tex}
 \end{document}
