+++ 07_initial_value_sensitivity.R
@@ -61,49 +61,95 @@ simulation_results <- bind_rows(bind_rows(sens_list) %>% mutate(link = 'logit'),
                                 bind_rows(sens_list_cloglog) %>% mutate(link = 'cloglog'))
 
 # reference parameter estimates from Candy 1991
-candy_cm_logit_pars = tibble(ref_estimate = c(5.49,5.49+3.90,5.49+6.74,5.49+10.21,5.49+15.77,5.49+21.76,-0.0457), parameter = c(paste('a', 1:6, sep = ''), 'b'))
+candy_cm_pars = tibble(ref_estimate = c(c(5.49,5.49+3.90,5.49+6.74,5.49+10.21,5.49+15.77,5.49+21.76,-0.0457),
+                                        c(3.32,3.32+2.53,3.32+4.40,3.32+6.71,3.32+10.14,3.32+14.20,-0.0307)),
+                       parameter = rep(c(paste('a', 1:6, sep = ''), 'b'),2),
+                       link = c(rep('logit', 7),rep('cloglog',7)))
 
 # calculate CV and correlation between start and estimate for converged estimates
-simulation_results %>% group_by(parameter, link) %>% summarize(range = paste(round(range(init)), collapse = ' - '), cv = sd(estimate, na.rm = TRUE) / mean(estimate, na.rm = TRUE), corr_init = cor(init, estimate, use = 'complete'), corr_p = cor.test(init, estimate, use = 'complete')$p.value)
-simulation_results %>% filter(value < 95 & link == 'logit' | value < 100 & link == 'cloglog') %>% group_by(parameter, link) %>% summarize(range = paste(round(range(init)), collapse = ' - '), cv = sd(estimate, na.rm = TRUE) / mean(estimate, na.rm = TRUE)*100, corr_init = cor(init, estimate, use = 'complete'), corr_p = cor.test(init, estimate, use = 'complete')$p.value)
+simulation_results %>% mutate(good = value < 95 & link == 'logit' | value < 100 & link == 'cloglog') %>% group_by(link) %>% summarize(finite_nll = sum(convergence == 0) / n(), good_nll_all = sum(good, na.rm = T)/n(), good_nll = sum(good, na.rm = T)/sum(convergence == 0))
+simulation_results  %>%
+  mutate(good = value < 95 & link == 'logit' | value < 100 & link == 'cloglog') %>%
+  group_by(parameter, link) %>% 
+  summarize(sim_range = paste(round(range(init),digits = 1), collapse = ' - '),
+            conv_range = paste(round(range(init[convergence == 0], na.rm = TRUE),digits = 1), collapse = ' - ')) %>% arrange(desc(link), parameter) -> results_unfiltered
+
+simulation_results %>% filter(value < 95 & link == 'logit' | value < 100 & link == 'cloglog') %>% group_by(parameter, link) %>% summarize(filt_range = paste(round(range(init),digits = 1), collapse = ' - '),cv = sd(estimate, na.rm = TRUE) / abs(mean(estimate, na.rm = TRUE))*100, corr_init = cor(init, estimate, use = 'complete'), corr_p = cor.test(init, estimate, use = 'complete')$p.value) %>% arrange(desc(link), parameter) -> results_filtered
+
+# assemble table for manuscript
+parameter_sens_table <- bind_cols(results_unfiltered, results_filtered[,3:6]) %>%
+  knitr::kable(digits = 3, format = 'latex', align = 'c', row.names = FALSE,
[…]
