### 2021-05-21 b1df113
manuscript: fix captions and add first draft of sensitivity analysis
