+++ outputs/parameter_sens_table.tex
@@ -1,0 +1,22 @@
+
+\begin{tabular}{cccccccc}
+\toprule
+Par. & Link & Sim. range & Conv. range & Filtered range & CV (\%) & $\rho^{o}$ & $P_\rho$\\
+\midrule
+a1 & logit & 1.3 - 20 & 1.3 - 20 & 1.3 - 20 & 0.692 & 0.101 & 0.115\\
+a2 & logit & 3.1 - 37.8 & 3.1 - 37.6 & 3.1 - 37.6 & 0.677 & 0.138 & 0.031\\
+a3 & logit & 5.4 - 54.1 & 5.4 - 54.1 & 5.4 - 54.1 & 0.627 & 0.123 & 0.055\\
+a4 & logit & 13 - 69.9 & 15.9 - 69.9 & 15.9 - 69.9 & 0.558 & 0.079 & 0.220\\
+a5 & logit & 22.9 - 85 & 22.9 - 85 & 22.9 - 85 & 0.548 & 0.006 & 0.929\\
+a6 & logit & 31.6 - 100.7 & 32.7 - 100.7 & 32.7 - 100.7 & 0.601 & -0.010 & 0.873\\
+b & logit & -1 - 1 & -1 - 0.8 & -1 - 0.8 & 0.570 & -0.106 & 0.097\\
+\addlinespace
+a1 & cloglog & 1.1 - 19.9 & 1.1 - 19.9 & 1.2 - 18.7 & 1.986 & 0.034 & 0.778\\
+a2 & cloglog & 3.3 - 38.9 & 4 - 37.3 & 4 - 34.4 & 1.774 & -0.045 & 0.709\\
+a3 & cloglog & 7.9 - 57 & 9.7 - 52.4 & 9.7 - 50.7 & 1.708 & -0.025 & 0.837\\
+a4 & cloglog & 15 - 67 & 16.3 - 66.9 & 16.3 - 63.8 & 1.689 & -0.055 & 0.651\\
+a5 & cloglog & 18.4 - 77.8 & 18.4 - 77.8 & 18.4 - 77.8 & 1.627 & 0.010 & 0.935\\
+a6 & cloglog & 22.8 - 91.6 & 30.7 - 91.6 & 30.7 - 83.2 & 1.663 & 0.039 & 0.751\\
+b & cloglog & -1 - 1 & -1 - 0.2 & -1 - 0.2 & 1.678 & -0.034 & 0.779\\
+\bottomrule
+\end{tabular}
