### 2021-05-21 97169c0
provide better starting values for cloglog_cm
