+++ outputs/cm_const_var_table.tex
@@ -9,7 +9,7 @@ $\alpha_1$ & $\alpha_2$ & $\alpha_3$ & $\alpha_4$ & $\alpha_5$ & $\alpha_6$ & $\
 5.47 & 9.36 & 12.21 & 15.67 & 21.22 & 27.19 & -0.05 & logit & R \verb+vglm+\\
 \addlinespace
 3.32 & 5.85 & 7.72 & 10.03 & 13.46 & 17.52 & -0.03 & cloglog & Original \citep{candy1991modeling}\\
-3.48 & 6.10 & 8.05 & 10.45 & 14.01 & 18.23 & -0.03 & cloglog & R \verb+optim+\\
+3.44 & 6.05 & 8.00 & 10.35 & 13.88 & 18.06 & -0.03 & cloglog & R \verb+optim+\\
 3.32 & 5.85 & 7.71 & 10.02 & 13.46 & 17.52 & 0.03 & cloglog & R \verb+clm+\\
 NA & NA & NA & NA & NA & NA & NA & cloglog & R \verb+vglm+\\
 \bottomrule
