+++ code/03_fit_dennis_et_al_model.R
@@ -1,6 +1,6 @@
 library(dplyr)
 library(kableExtra)
-budworm_counts <- readr::read_csv('data/budworm_counts.csv')
+budworm_counts <- readr::read_csv('data/budworm_counts.csv', col_types = "dddd")
 
 #optimize likelihood directly rather than doing the IRLS approach
 nll_cm_dennis <- function(par, count, total, stage, ddeg){
@@ -25,7 +25,7 @@ nll_cm_dennis <- function(par, count, total, stage, ddeg){
   nll <- -1*sum(count*log(pred))
   return(nll)
 }
-logit_cm_dennis_nll <- optim(par = c(A1 = 150, A2 = 230, A3 = 280, A4 = 330, A5 = 440, A6 = 580, BB = 3), nll_cm_dennis, count = budworm_counts$count, total = budworm_counts$total, stage = budworm_counts$stage, ddeg = budworm_counts$ddeg, hessian = TRUE, control = list(trace=1), method = 'BFGS')
+logit_cm_dennis_nll <- optim(par = c(A1 = 150, A2 = 230, A3 = 280, A4 = 330, A5 = 440, A6 = 580, BB = 3), nll_cm_dennis, count = budworm_counts$count, total = budworm_counts$total, stage = budworm_counts$stage, ddeg = budworm_counts$ddeg, hessian = TRUE, control = list(trace=0), method = 'BFGS')
 #logit_cm_dennis_nll <- add_se_vcov_nll_hessian(logit_cm_dennis_nll)
 
 predicted_proportion <- function(par, stage, ddeg){
