### 2021-05-21 6ca88b1
tidy appendix captions
