+++ article/appendix.tex
@@ -22,18 +22,17 @@ The optimisation of the cloglog model was much less stable. Only 54\% of simulat
 \begin{table}[bph]
   \small
     \centering
-    \caption{Summary statistics of parameter sensitivities to starting values for the cumulative model with constant variance (Eqn.~\ref{eq:candy_cm_count_form}). 
+    \caption{Summary statistics of parameter sensitivities to starting values for the cumulative model with constant variance (Eqn.~\ref{eq:candy_cm_count_form}). The coefficients of variation (CV), correlation coefficients between initial and convergend values $\rho^{\circ}$, and their corresponding p-values $P_\rho$ were calculated for the filtered parameter estimates only.
     }
   \input{../outputs/parameter_sens_table}
   \label{tab:tabS2}
 \end{table}
 
-The ranges of initial values, as well as  procedure resulted in initial value ranges of 
 
 \begin{figure}[htbp]
   \centering
   \includegraphics[width=\textwidth]{../figures/fig3_initial_value_sensitivity_unfiltered.pdf}
-  \caption{Optimisation results for randomly drawn initial values for the cumulative model with constant variance. Parameter estimates are coloured by the log of the negative log-likelihood of the corresponding model fit. Yellow broken lines correspond to parameter values reported in \citep{candy1991modeling}. The majority of initial values resulted in estimates close to the original publication for the logit model, but convergence to local optima far from the published parameters was common in the cloglog link model.}
+  \caption{Optimisation results for randomly drawn initial values for the cumulative model with constant variance. Parameter estimates are coloured by the log of the negative log-likelihood of the corresponding model fit. Yellow broken lines correspond to parameter values reported in \citep{candy1991modeling}. The majority of initial values resulted in estimates close to the original publication for the logit model (bottom), but convergence to local optima far from the published parameters was common in the cloglog link model (top).}
   \label{fig:fig3}
 \end{figure} 
 
