### 2021-05-21 pboesu/replication_candy_1991#6 @pboesu
Most of the original discrepancies between parameter estimates could be resolved by adding parameter scaling factors to the optimisation routine. Regardless I retained the initial value sensitivity analysis example added in a9cc8643d35fc89e6e5921d71b5e2a11555b82e9 , b1df113459f834b71fc94bbec14abaa8c3cf5625
