### 2021-05-21 pboesu/replication_candy_1991#5 @pboesu
Revision checklist
- [x]  Expanded discussion of numerical limitations when transforming the sequential model results to probability scale added in 94e71ff5a84972b32fd5ce2f464d9f9f0f7e7a53
