### 2021-05-24 94e71ff
more details on numerical aspects of optimisation and fitted probabilities
