+++ outputs/cm_const_var_table.tex
@@ -4,12 +4,12 @@
 Method & $\alpha_1$ & $\alpha_2$ & $\alpha_3$ & $\alpha_4$ & $\alpha_5$ & $\alpha_6$ & $\beta$ & Link\\
 \midrule
 Original \citep{candy1991modeling} & 5.49 & 9.39 & 12.23 & 15.70 & 21.26 & 27.25 & -0.05 & logit\\
-R \verb+optim+ & 5.53 & 9.48 & 12.35 & 15.86 & 21.46 & 27.52 & -0.05 & logit\\
+R \verb+optim+ & 5.47 & 9.38 & 12.22 & 15.69 & 21.25 & 27.22 & -0.05 & logit\\
 R \verb+clm+ & 5.47 & 9.36 & 12.21 & 15.67 & 21.22 & 27.19 & 0.05 & logit\\
 R \verb+vglm+ & 5.47 & 9.36 & 12.21 & 15.67 & 21.22 & 27.19 & -0.05 & logit\\
 \addlinespace
 Original \citep{candy1991modeling} & 3.32 & 5.85 & 7.72 & 10.03 & 13.46 & 17.52 & -0.03 & cloglog\\
-R \verb+optim+ & 3.44 & 6.05 & 8.00 & 10.35 & 13.88 & 18.06 & -0.03 & cloglog\\
+R \verb+optim+ & 3.35 & 5.90 & 7.79 & 10.12 & 13.57 & 17.67 & -0.03 & cloglog\\
 R \verb+clm+ & 3.32 & 5.85 & 7.71 & 10.02 & 13.46 & 17.52 & 0.03 & cloglog\\
 R \verb+vglm+ & NA & NA & NA & NA & NA & NA & NA & cloglog\\
 \bottomrule
