### 2021-05-24 pboesu/replication_candy_1991#3 @pboesu
Revisions checklist:

- [x]  clarify that only intermediate outputs are stored as rds to minimize typecasting and rounding errors c0d04e517e422da758edbc85f6c0aaf91a81c365
- [x]  improve readability of comment lines
- [x]  split function creation from function use
- [x]  minimize uninformative messages - silenced readr and dplyr messages, and model fitting trace reports
- [x]  explicitly encourage stepping through scripts to follow the reproduction
