### 2021-05-24 640afb7
update .Rprofile and README to give pointers about avoiding renv
