### 2021-05-24 bef2c98
move function definitions to separate script; tidy comments; avoid dplyr and ggplot warnings by explicitly addressing groups and NA handling
