+++ code/02_fit_candy_1991_models.R
@@ -1,52 +1,20 @@
 library(dplyr)
-#Poisson deviance fitting approach
-#
 # read data ddeg day-degrees, total total individuals in stage,  
 #! stage stage number 1-5 instars, 6 pupa, 7 adult
 #! count total individuals in stage
 budworm_counts <- readr::read_csv('data/budworm_counts.csv', col_types = "dddd")
-budworm_counts$occ <- as.numeric(as.factor(budworm_counts$ddeg)) #encode sampling occasion
-budworm_counts$alpha_index = budworm_counts$stage + 1 #encode an index variable to facilitate fitting the linear predictor
+budworm_counts$occ <- as.numeric(as.factor(budworm_counts$ddeg)) # encode sampling occasion
+budworm_counts$alpha_index = budworm_counts$stage + 1 # encode an index variable to facilitate fitting the linear predictor
 budworm_counts <- group_by(budworm_counts, ddeg) %>%
   arrange(ddeg,stage) %>% 
-  mutate(cumulative = cumsum(count)) %>% #calculate cumulative counts across stages for cumulative likelihood
-  mutate(n_star = ifelse(stage ==1, total, total - lag(cumulative)), #calculate N*_ij for GLM estimation of sequential model
+  mutate(cumulative = cumsum(count)) %>% # calculate cumulative counts across stages for cumulative likelihood
+  mutate(n_star = ifelse(stage ==1, total, total - lag(cumulative)), # calculate N*_ij for GLM estimation of sequential model
          stage_factor = as.factor(stage)) 
 
-#likelihood function for cumulative model with constant variance
-poisson_nll_cm_candy <- function(par, data, linkinv = gtools::inv.logit){
-  alpha <- c(-10, par[1:6], 5000)
-  beta = par[7]
-  pred_n <- rep(NA, nrow(data))
-  for (i in 1:nrow(data)){
-    pred_n[i] <- with(data,
-                      total[i] * (linkinv(alpha[alpha_index[i]] + beta*(ddeg[i])) - linkinv(alpha[alpha_index[i]-1] + beta*(ddeg[i])))
-    )
-  }
-  #avoid numerical underflow during fitting
-  pred_n <- pmax(pred_n, .Machine$double.xmin)
-  nll = sum(-1*dpois(data$count, pred_n, log=T))
-  return(nll)
-}
 
-#likelihood function for cumulative model with proportional variance
-poisson_nll_cm_candy_dennis <- function(par, data){
-  alpha <- c(-10, par[1:6], 5000)
-  beta = par[7]
-  pred_n <- rep(NA, nrow(data))
-  for (i in 1:nrow(data)){
-    pred_n[i] <- with(data,
-                             total[i]*((exp(alpha[alpha_index[i]]/sqrt(ddeg[i]) + beta*sqrt(ddeg[i]))/(1 + exp(alpha[alpha_index[i]]/sqrt(ddeg[i]) + beta*sqrt(ddeg[i])))) - 
-                                         (exp(alpha[alpha_index[i]-1]/sqrt(ddeg[i]) + beta*sqrt(ddeg[i]))/(1 + exp(alpha[alpha_index[i]-1]/sqrt(ddeg[i]) + beta*sqrt(ddeg[i])))))
-    )}
-  #avoid numerical underflow during fitting
-  pred_n <- pmax(pred_n, .Machine$double.xmin)
-  nll = -1*sum(dpois(data$count, pred_n, log=T))
-  return(nll)
-}
 
-#get parameter estimates for cumulative models
-#use optimal starting values from supplementary simulations
[…]
