+++ code/00_function_definitions.R
@@ -1,0 +1,147 @@
+# function definitions
+
+# likelihood function for cumulative model with constant variance
+poisson_nll_cm_candy <- function(par, data, linkinv = gtools::inv.logit){
+  alpha <- c(-10, par[1:6], 5000)
+  beta = par[7]
+  pred_n <- rep(NA, nrow(data))
+  for (i in 1:nrow(data)){
+    pred_n[i] <- with(data,
+                      total[i] * (linkinv(alpha[alpha_index[i]] + beta*(ddeg[i])) - linkinv(alpha[alpha_index[i]-1] + beta*(ddeg[i])))
+    )
+  }
+  # avoid numerical underflow during fitting
+  pred_n <- pmax(pred_n, .Machine$double.xmin)
+  nll = sum(-1*dpois(data$count, pred_n, log=T))
+  return(nll)
+}
+
+# likelihood function for cumulative model with proportional variance
+poisson_nll_cm_candy_dennis <- function(par, data){
+  alpha <- c(-10, par[1:6], 5000)
+  beta = par[7]
+  pred_n <- rep(NA, nrow(data))
+  for (i in 1:nrow(data)){
+    pred_n[i] <- with(data,
+                      total[i]*((exp(alpha[alpha_index[i]]/sqrt(ddeg[i]) + beta*sqrt(ddeg[i]))/(1 + exp(alpha[alpha_index[i]]/sqrt(ddeg[i]) + beta*sqrt(ddeg[i])))) - 
+                                  (exp(alpha[alpha_index[i]-1]/sqrt(ddeg[i]) + beta*sqrt(ddeg[i]))/(1 + exp(alpha[alpha_index[i]-1]/sqrt(ddeg[i]) + beta*sqrt(ddeg[i])))))
+    )}
+  # avoid numerical underflow during fitting
+  pred_n <- pmax(pred_n, .Machine$double.xmin)
+  nll = -1*sum(dpois(data$count, pred_n, log=T))
+  return(nll)
+}
+
+# likelihood function for the cumulative model with constant variance without numerical thresholds against underflow 
+poisson_nll_cm_candy_inf <- function(par, data, linkinv = gtools::inv.logit){
+  alpha <- c(-10, par[1:6], 5000)
+  beta = par[7]
+  pred_n <- rep(NA, nrow(data))
+  for (i in 1:nrow(data)){
+    pred_n[i] <- with(data,
+                      total[i] * (linkinv(alpha[alpha_index[i]] + beta*(ddeg[i])) - linkinv(alpha[alpha_index[i]-1] + beta*(ddeg[i])))
+    )
+  }
+  nll = sum(-1*dpois(data$count, pred_n, log=T))
+  return(nll)
+}
+
+# function to repeatedly fit the cumulative logit model with randomised starting values
+cm_cand_sens <- function(iter, progress = TRUE, linkinv = gtools::inv.logit){
+  if (progress == TRUE & (iter %% 10 == 0)) print(paste(Sys.time(), '- iteration', iter))
+  
+  # draw random initial values
+  alpha_start <- cumsum(runif(6,1,20))
+  beta_start <- runif(1, -1,1)
+  # estimate model parameters
[…]

[…]
