### 2021-05-24 c0d04e5
leave scripts as standalone blocks and encourage user to step through them; comment out failing VGAM code
