+++ code/04_fit_VGAM_models.R
@@ -1,9 +1,14 @@
+# fit sequential and cumulative models using VGAM
+# Note that neither of the `vglm` models converged when using the cloglog link.
+# This behaviour has been reported to the package maintainer as a potential bug.
+# The model specifications are left in the code but commented out.
+
 library(VGAM)
 budworm_table <- readr::read_csv('data/budworm_table.csv', col_types = 'dddddddddd')
 
-# fit cumulative model
-budworm_cumulative_cloglog_vglm <- try(vglm(cbind(stage1, stage2, stage3, stage4, stage5, stage6, stage7) ~ ddeg,
-                                        cumulative(reverse = FALSE, link = clogloglink(bvalue = 1e-8), parallel = TRUE), coefstart = c(3,6,8,10,14,17,-0.04), data = budworm_table))# fails
+# fitting the  cumulative model with cloglog link fails - this has been reported to the package maintainer
+# budworm_cumulative_cloglog_vglm <- try(vglm(cbind(stage1, stage2, stage3, stage4, stage5, stage6, stage7) ~ ddeg,
+#                                         cumulative(reverse = FALSE, link = clogloglink(bvalue = 1e-8), parallel = TRUE), coefstart = c(3,6,8,10,14,17,-0.04), data = budworm_table))# fails
 
 
 budworm_cumulative_logit_vglm <- vglm(cbind(stage1, stage2, stage3, stage4, stage5, stage6, stage7) ~ ddeg,
@@ -22,10 +27,11 @@ saveRDS(candy_vglm_cm_estimates, "outputs/candy_vglm_cm_estimates.RDS")
 
 # fit sequential model
 budworm_sratio_logit_vglm <- vglm(cbind(stage1, stage2, stage3, stage4, stage5, stage6, stage7) ~ ddeg,
-                                  sratio(link = "logitlink", parallel = FALSE), data = budworm_table)
+                                  sratio(link = "logitlink", parallel = FALSE), data = budworm_table)# this function call will succeed but produce warnings about replacing working weights to prevent numerical underflow.
 
-budworm_sratio_cloglog_vglm <- try(vglm(cbind(stage1, stage2, stage3, stage4, stage5, stage6, stage7) ~ ddeg_cent,
-                                        sratio(link = "clogloglink", parallel = FALSE), data = budworm_table, control = vglm.control(checkwz = TRUE, wzepsilon = 0.5)))# sratio model with clogloglink fails because of some numerical issue - but check the expected binomial sample sizes. Candy writes "Note that the n_ij for which the binomial sample size is N^*_ij is zero must be excluded from the fit by by giving them a prior weight of zero. VGAM requires positive weights but suggests using weights "such as 1e-8" to effectively exclude observations"
+# fitting the sratio model with clogloglink fails - this has been reported to the package maintainer
+# budworm_sratio_cloglog_vglm <- try(vglm(cbind(stage1, stage2, stage3, stage4, stage5, stage6, stage7) ~ ddeg_cent,
+#                                        sratio(link = "clogloglink", parallel = FALSE), data = budworm_table, control = vglm.control(checkwz = TRUE)))#
 
 # save parameter estimates
 candy_vglm_sm_estimates <- as.data.frame(rbind(unname(coef(budworm_sratio_logit_vglm)[1:6]),unname(coef(budworm_sratio_logit_vglm)[7:12]),rep(NA,6),rep(NA,6)))
