### 2021-05-24 e16a634
add informations about warnings to source code
