+++ article/appendix.tex
@@ -16,9 +16,9 @@ Optimisation proceeded using the likelihood outlined in with a numerical thresho
 Parameters were further filtered to remove parameter sets where convergence was achieved at a finite likelihood, but with a substantially larger overall log-likelihood than the best model fit. Thresholds for this were determined empirically and set at 93 for the logit link and 100 for the cloglog model.
 
 \subsection{Results}
-The optimisation was reasonably robust for the logit link model. 82\% of simulations achieved convergence and of these over 98\% met the filtering threshold (Figure~\ref{fig:fig3}). Among the filtered parameter sets variability in the final estimates was small (parameter-wise coefficients of variation were all smaller than 0.5\%) and there was no correlation between initial values and parameter estimates (Table~\ref{tab:tabS2}). Convergence failed for initial values $\beta^{\circ}>0.8$, and convergence failures were likely, but not guaranteed for $\beta^{\circ}\approx 0$ (Figure~\ref{fig:fig4}).
+The optimisation was reasonably robust for the logit link model. 82\% of simulations achieved convergence and of these over 98\% met the filtering threshold (Figure~\ref{fig:figS1}). Among the filtered parameter sets variability in the final estimates was small (parameter-wise coefficients of variation were all smaller than 0.5\%) and there was no correlation between initial values and parameter estimates (Table~\ref{tab:tabS2}). Convergence failed for initial values $\beta^{\circ}>0.8$, and convergence failures were likely, but not guaranteed for $\beta^{\circ}\approx 0$ (Figure~\ref{fig:figS2}).
 
-The optimisation of the cloglog model was much less stable. Only 54\% of simulations resulted in convergence, and of those fewer than half (43\%) met the filtering threshold. This is apparent as a large cluster of estimates deviating from the reference estimates in Figure~\ref{fig:fig3}. Variability in the filtered parameter estimates was larger than for the logit model, but still small in absolute terms (parameter-wise coefficients of variation were all between 1.6\% and 2.0\%) and there was no correlation between initial values and parameter estimates (Table~\ref{tab:tabS2}). Convergence failed for most $\beta^{\circ}>0$, and generally for $\beta^{\circ}>0.2$ (Figure~\ref{fig:fig5}). No general pattern was apparent for initial value combinations that led to the model converging at local optima far from the optimum associated with the reference parameter values.
+The optimisation of the cloglog model was much less stable. Only 54\% of simulations resulted in convergence, and of those fewer than half (43\%) met the filtering threshold. This is apparent as a large cluster of estimates deviating from the reference estimates in Figure~\ref{fig:figS1}. Variability in the filtered parameter estimates was larger than for the logit model, but still small in absolute terms (parameter-wise coefficients of variation were all between 1.6\% and 2.0\%) and there was no correlation between initial values and parameter estimates (Table~\ref{tab:tabS2}). Convergence failed for most $\beta^{\circ}>0$, and generally for $\beta^{\circ}>0.2$ (Figure~\ref{fig:figS3}). No general pattern was apparent for initial value combinations that led to the model converging at local optima far from the optimum associated with the reference parameter values.
 
 
 \begin{table}[bph]
@@ -35,19 +35,19 @@ The optimisation of the cloglog model was much less stable. Only 54\% of simulat
   \centering
   \includegraphics[width=\textwidth]{../figures/figS1_initial_value_sensitivity_unfiltered.pdf}
[…]
