### 2021-05-27 49b525e
revisions to text following the implementation of parscale throughout & proof reading
