+++ article/content.tex
@@ -5,7 +5,7 @@ It defines the timing and duration of growth and reproduction and thereby determ
 Phenological analyses often focus on the timing of particular events, such as the dates of peak caterpillar abundance \citep{shutt2019spatial}.
 However, for many biological phenomena exact dates of particular events are more difficult to observe than the state of the system itself.
 For example, repeated but sparse survey visits may record whether a plant is in bud, flowering, or setting fruit, but not the exact dates when each of those stages was reached.
-Such observations can be used to categorize an organism's state into discrete classes which usually follow natural ordering, e.g. from least to most developed. 
+Such observations can be used to categorize an organism's state into discrete classes which usually follow natural ordering, e.g.\ from least to most developed. 
 The resulting data can be described using ordinal regression models \cite{mccullagh1980regression,agresti2010analysis} which then allow inferences about phenological progression.
 
 I here replicate a number of ordinal regression models that were employed by \citet{dennis1986stochastic} and \citet{candy1991modeling} to describe insect phenology. 
@@ -70,9 +70,9 @@ For the purpose of parameter estimation \citet{candy1991modeling} re-expressed t
 \end{equation}
 and fits it using a Poisson likelihood \citep{thompson1981composite}. 
 No code or initial values for the likelihood optimisation are provided for this estimation procedure in the original paper.   
-I therefore created an R version of the estimation procedure which directly optimizes a Poisson log-likelihood for Equation~\ref{eq:candy_cm_count_form} using the BFGS method in the \verb+optim+ function.
+I therefore created an R version of the estimation procedure which directly optimizes a Poisson log-likelihood for Equation~\ref{eq:candy_cm_count_form} using the BFGS method in the \verb+optim+ function and a parameter scaling value of 0.01 for $\beta$ relative to the cut-point parameters..
 Initial values for the optimisation were determined using a random search across plausible start values (see Appendix~\ref{sec:appendix}).
- The cumulative model is implemented in various R packages, including in the \verb+vglm+ function in \verb+VGAM+ \citep{VGAM} and the \verb+clm+ function in \verb+ordinal+ \cite{ordinal} and for comparison I also attempted to fit the models using these functions using their default settings.
+ The cumulative model is implemented in various R packages, including in the \verb+vglm+ function in \verb+VGAM+ \citep{VGAM} and the \verb+clm+ function in \verb+ordinal+ \cite{ordinal} and for comparison I also attempted to fit the models using these functions with their default settings.
 
 \subsection{Cumulative model with proportional variance}
 \citet{dennis1986stochastic} used a different parameterisation of the ordinal model with a logit link, i.e. assuming a logistic distribution for $S(t)$, such that the probability that an insect's development at time $t$ has not exceeded $s$ amounts to 
@@ -82,7 +82,7 @@ Pr(S(t) \leq s) = \left\{ 1 + \exp\left[-\left(\frac{s-t}{\sqrt{b^2t}}\right)\ri
 \end{equation}
 where $b^2$ is a positive constant. 
 The cumulative distribution function in Eqn.~\ref{eqn:cumpropcdf} corresponds to a logistic distribution with a mean of $t$ and a variance which increases proportional to the mean as $(\pi^2/3)b^2t$.
-At any fixed time $t$ the thresholds $a_j$ segment the probability distribution function in to $r$ parts and the area under the curve between $a_{j-1}$ and $a_j$ gives the probability that the insect will be in stage $j$ at time $t$.
+At any fixed time $t$ the thresholds $a_j$ segment the probability distribution function into $r$ parts and the area under the curve between $a_{j-1}$ and $a_j$ gives the probability that the insect will be in stage $j$ at time $t$.
 
[…]
