+++ article/appendix.tex
@@ -6,25 +6,25 @@
 \renewcommand\thetable{S\arabic{table}}    
 \setcounter{table}{0}
 
-\section{Appendix: Sensitivity of the optimisation procedure to intial values}
+\section{Appendix: Sensitivity of the optimisation procedure to initial values}
 \subsection{Approach}
 \label{sec:appendix}
 The sensitivity of parameter estimates for the cumulative model with constant variance (Eqn.~\ref{eq:candy_cm_count_form}) obtained using the direct likelihood optimisation with \verb+optim+ was assessed by simulation.
-Simulations consisted of drawing a set of random initial values. To preserve the ordering of the cut-point parameters $\alpha_i$, their initial values $\alpha^{\circ}_i$ were assembled as the cumulative sum of six independent draws $$x_j \sim \mathrm{Uniform}(1,20),$$ i.e. $$\alpha^{\circ}_i = \sum_{j=1}^i x_j.$$ The initial value for $\beta$ was drawn as $$\beta^{\circ}\sim\mathrm{Uniform}(-1,1).$$
+Simulations consisted of drawing a set of random initial values. To preserve the ordering of the cut-point parameters $\alpha_i$, their initial values $\alpha^{\circ}_i$ were assembled as the cumulative sum of six independent draws $$x_j \sim \mathrm{Uniform}(1,20), \qquad j = 1,\ldots, 6$$ i.e. $$\alpha^{\circ}_i = \sum_{j=1}^i x_j.$$ The initial value for $\beta$ was drawn as $$\beta^{\circ}\sim\mathrm{Uniform}(-1,1).$$
 
-Optimisation proceeded using the likelihood outlined in Section~\ref{sec:const_var_cm} with a numerical threshold to prevent the Poisson likelihood from numerical underflowing. As this allowed convergence at an infinite likelihood, the likelihood was evaluated one final time without thresholding at the converged parameter values and parameters were only retained when the unthresholded likelihood was finite. 
+Optimisation proceeded using the likelihood outlined in Section~\ref{sec:const_var_cm} with a numerical threshold at the machine's precision to prevent the Poisson likelihood from approaching negative infinity. The optimisation routine used the BFGS method and parameter scaling value for $\beta$ of 0.01 relative to the cut-point parameters $\alpha_i$. This improved overall convergence rates but allowed convergence at an infinite likelihood, the likelihood was therefore evaluated one final time without thresholding at the converged parameter values and parameters were only retained when the unthresholded likelihood was finite. 
 Parameters were further filtered to remove parameter sets where convergence was achieved at a finite likelihood, but with a substantially larger overall log-likelihood than the best model fit. Thresholds for this were determined empirically and set at 93 for the logit link and 100 for the cloglog model.
 
 \subsection{Results}
-The optimisation was reasonably robust for the logit link model. 82\% of simulations achieved convergence and of these over 98\% met the filtering threshold (Figure~\ref{fig:figS1}). Among the filtered parameter sets variability in the final estimates was small (parameter-wise coefficients of variation were all smaller than 0.5\%) and there was no correlation between initial values and parameter estimates (Table~\ref{tab:tabS2}). Convergence failed for initial values $\beta^{\circ}>0.8$, and convergence failures were likely, but not guaranteed for $\beta^{\circ}\approx 0$ (Figure~\ref{fig:figS2}).
[…]
