### 2021-05-28 9706b87
add parscale to optim calls and Bonferroni correction to correlation p-values; rerun replications throughout
