+++ code/00_function_definitions.R
@@ -54,17 +54,59 @@ cm_cand_sens <- function(iter, progress = TRUE, linkinv = gtools::inv.logit){
   alpha_start <- cumsum(runif(6,1,20))
   beta_start <- runif(1, -1,1)
   # estimate model parameters
-  cm_candy_nll <- optim(par = c(alpha_start,beta_start), poisson_nll_cm_candy, data = budworm_counts, hessian = TRUE, control = list(trace=0), method = 'BFGS', linkinv = linkinv)
+  cm_candy_nll <- optim(par = c(alpha_start,beta_start), poisson_nll_cm_candy, data = budworm_counts, hessian = TRUE, control = list(trace=0, parscale = c(rep(1,6), 0.01)), method = 'BFGS', linkinv = linkinv)
   # check that converged parameters actually have a finite likelihood
   if (is.finite(poisson_nll_cm_candy_inf(par = cm_candy_nll$par, data = budworm_counts, linkinv = linkinv))){
-    sens_out <- tibble(init = c(alpha_start,beta_start),
+    sens_out <- tibble::tibble(init = c(alpha_start,beta_start),
                        estimate = cm_candy_nll$par,
                        parameter = c(paste('a', 1:6, sep = ''), 'b'),
                        convergence = cm_candy_nll$convergence,
                        iter = iter,
                        value = cm_candy_nll$value) 
   } else { # else return an indicator of false convergence
-    sens_out <- tibble(init = c(alpha_start,beta_start),
+    sens_out <- tibble::tibble(init = c(alpha_start,beta_start),
+                       convergence = -1,
+                       parameter = c(paste('a', 1:6, sep = ''), 'b'),
+                       iter = iter)
+  }
+  return(sens_out)
+}
+
+# function to draw the likelihood surface for the cumulative model with constant variance 
+cm_nll_surface <- function(iter, progress = TRUE, linkinv = gtools::inv.logit){
+  if (progress == TRUE & (iter %% 1000 == 0)) print(paste(Sys.time(), '- iteration', iter))
+  # draw random initial values
+  alpha_start <- cumsum(c(runif(1,0,9),runif(5,0.1,7)))
+  beta_start <- runif(1, -0.1,0)
+  # estimate model parameters
+  nll = poisson_nll_cm_candy(par = c(alpha_start,beta_start), data = budworm_counts, linkinv = linkinv)
+  # check that converged parameters actually have a finite likelihood
+  sens_out <- tibble::tibble(init = c(alpha_start,beta_start),
+                             parameter = c(paste('a', 1:6, sep = ''), 'b'),
+                             iter = iter,
+                             value = nll)
+  return(sens_out)
+}
+
+# function to repeatedly fit the VGAM cumulative logit model with randomised starting values
+cm_vgam_sens <- function(iter, progress = TRUE, link = clogloglink()){
+  if (progress == TRUE & (iter %% 200 == 0)) print(paste(Sys.time(), '- iteration', iter))
+  
+  # draw random initial values
+  alpha_start <- cumsum(runif(6,0.1,5))
+  beta_start <- runif(1, -0.1,-0.01)
+  # estimate model parameters
+  cm_vgam_nll <- try(vglm(cbind(stage1, stage2, stage3, stage4, stage5, stage6, stage7) ~ ddeg_cent, cumulative(reverse = FALSE, link = clogloglink(), parallel = TRUE), coefstart = c(alpha_start, beta_start), data = budworm_table, control = vglm.control(checkwz = TRUE)))#
+  # check that model didn't error out
+  if (!inherits(cm_vgam_nll, 'try-error')){
+    sens_out <- tibble::tibble(init = c(alpha_start,beta_start),
+                       estimate = coef(cm_vgam_nll),
[…]
