### 2021-05-28 pboesu/replication_candy_1991#4 @pboesu
Correspondence with the maintainer of VGAM
