### 2021-05-28 pboesu/replication_candy_1991#2 @pboesu
- [x] added more explicit language around the use/avoidance of `renv` in both README and .Rprofile 640afb73131863aa944f8a21cd4b98aa6aa01a60
- [x] provided `sessionInfo()` in README 640afb73131863aa944f8a21cd4b98aa6aa01a60
