### 2021-05-28 68bf3e8
additional clarification about non-renv usage
