+++ article/bibliography.bib
@@ -27,7 +27,7 @@
 	Date-Modified = {2020-08-28 17:59:33 +0100},
 	Doi = {10.1177/2515245918823199},
 	Journal = {Advances in Methods and Practices in Psychological Science},
-	Month = {feb},
+	Month = {2},
 	Number = {1},
 	Pages = {77--101},
 	Publisher = {{SAGE} Publications},
@@ -42,7 +42,7 @@
 	Date-Added = {2020-08-28 17:58:15 +0100},
 	Date-Modified = {2020-08-28 17:58:17 +0100},
 	Doi = {10.1002/9780470594001},
-	Month = {mar},
+	Month = {3},
 	Publisher = {John Wiley {\&} Sons, Inc.},
 	Title = {Analysis of Ordinal Categorical Data},
 	Year = 2010,
@@ -55,7 +55,7 @@
 	Date-Modified = {2020-08-28 17:52:55 +0100},
 	Doi = {10.1111/j.2517-6161.1980.tb01109.x},
 	Journal = {Journal of the Royal Statistical Society: Series B (Methodological)},
-	Month = {jan},
+	Month = {1},
 	Number = {2},
 	Pages = {109--127},
 	Publisher = {Wiley},
@@ -71,7 +71,7 @@
 	Date-Modified = {2020-08-28 17:25:06 +0100},
 	Doi = {10.1093/ee/15.3.547},
 	Journal = {Environmental Entomology},
-	Month = {jun},
+	Month = {6},
 	Number = {3},
 	Pages = {547--554},
 	Publisher = {Oxford University Press ({OUP})},
@@ -87,7 +87,7 @@
 	Date-Modified = {2020-08-28 17:21:22 +0100},
 	Doi = {10.1093/ee/15.3.540},
 	Journal = {Environmental Entomology},
-	Month = {jun},
+	Month = {6},
 	Number = {3},
 	Pages = {540--546},
 	Publisher = {Oxford University Press ({OUP})},
