### 2021-05-28 #46 @pboesu
Ok! Revisions are done! - Thanks again to @aammd and @juniperlsimonis for their constructive comments which encouraged me to dive into the optimisation code much deeper and learn a thing or two along the way!

The overall diff can be seen here: https://github.com/pboesu/replication_candy_1991/compare/v0.1...v0.2

I apologize that the git history is a bit convoluted. Halfway through my revisions, I discovered that my failure to use parameter scaling in the optimisation procedure had created some of the numerical issues I discussed in the original submission and which were commented on by the reviewers.
Adding parameter scaling resolved several of these numerical instabilities, so there is a set of responses in the git history which were ultimately superceded, but for the sake of completeness I haven't squashed those commits.

I have provided summaries of changes in each of the issues in the replication repository. And a high-level response to each issue is given below

1. [Review by aammd](https://github.com/pboesu/replication_candy_1991/issues/1)
I addressed all but one issues raised by @aammd. This included adding box constraints to one set of optimisations to avoid error messages during the optimisation run, investigating the sensitivity of parameter estimates to using a literal implementation of the inverse logit link versus the corresponding `plogis` function, and some textual/typesetting issues. I did not have a good idea about making the reformatted data set more discoverable than it will be following the publication and archival of the replication paper. As it is, the data are provided in clear text in the original publication (albeit in an 'untidy' format), and I do not think that a separate re-publication of the data beyond the replication paper is justified given the small size of the data set.

2. [Installation issues](https://github.com/pboesu/replication_candy_1991/issues/2)
I responded to these suggestions by providing more explicit information to the user about using or not using `renv` for dependenccy management, both in [README](https://github.com/pboesu/replication_candy_1991/compare/v0.1..v0.2#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5) and in [.Rprofile](https://github.com/pboesu/replication_candy_1991/compare/v0.1..v0.2#diff-bd3c4db940418433876cd2f7ba7ab9a320166be4a13a1a29bd169d3a33d4e5ec). Dependency management in R is a mess to be perfectly honest, but for all of it's imperfections and peculiarities `renv` does provide a standardized way to simplify it, even if that comes at the cost of a non-standard file structure (i.e. the repository-specific .Rprofile which is separate from other dependency / function loading code in the project). I hope the added comments and `sessionInfo()` details give the advanced R user sufficient information to make an informed choice in how to approach recreating the runtime environment.

3. [Code structuring](https://github.com/pboesu/replication_candy_1991/issues/3)
Again, I chose a compromise here. I split out function definitions from function use, but otherwise left the structure of the analysis code as is - as sequentially numbered R scripts, that can be run stand-alone if desired. I clarified the role of using binary `.rds` files as intermediate storage, and minimized the messages that are spit out to the console when running the code. I now explicitly encourage the user to step through the scripts to follow the reproduction, although simply sourcing them in order will still reproduce the manuscript without any need for user interaction.

4. [VGAM errors](https://github.com/pboesu/replication_candy_1991/issues/4)
This is the only major issue that I could not resolve. Model fits with VGAM for cloglog link models continue to either error out, or result in false convergence. I have been in communication with the maintainer of the VGAM package, Thomas Yee, and we established that the underlying cause for this is that the models are estimating parameters at the boundary of the parameter space (i.e. cut-point parameters that are very close to 1 on the probability scale).  For the cumulative family Thomas managed to find an initial fix which removes the originally reported error, but the model continues to converge at local optima far from the solution found with the other algorithms. Thomas is keen to further investigate this issue but is unlikely to do so in the near future due to other commitments.  I have commented out all code using `VGAM::vglm` for cloglog link to avoid throwing error messages when running the replication and added language to the manuscript to highlight that this remains an open issue.

5. [Warnings related to probabilities close to 1](https://github.com/pboesu/replication_candy_1991/issues/5)
I have added more details about the data to the manuscript and a description of how the nature of the data (many 0 cell counts) lead to extreme parameter estimates and hence the warnings produced by `stats::glm` when fitting the corresponding models.

6. [Optim sensitivity](https://github.com/pboesu/replication_candy_1991/issues/6)
I have added an appendix which investigates the behaviour of the direct likelihood optimisation under random initial value proposals. This led me to discover that many of the original instabilities (i.e. parameter variation on the order of 1-2% for different initial values) were a result of not appropriately specifying parameter scaling values in `optim`.  I decided to retain the appendix regardless, as I believe it still has some useful results showing that the cloglog link model is intrinsically less numerically stable than the logit link one.

7. [Figures](https://github.com/pboesu/replication_candy_1991/issues/7)
All suggestions in this issue were adressed.

8. [Manuscript text](https://github.com/pboesu/replication_candy_1991/issues/8)
I added details about function arguments and specific algorithms for the fitting functions and qunatified differences in parameter estimates where relevant.
