### 2021-06-03 #46 @tpoisot
I am going to assume that the :tada: and :rocket: emojis on [the response[(https://github.com/ReScience/submissions/issues/46#issuecomment-850565625) means that @aammd and @juniperlsimonis are happy with the revision, and so this is an **accept** for me.

@rougier - I might need a refresher with the publishing pipeline
