### 2021-06-07 #46 @rougier
@tpoisot I can publish it for you, might be easier.  @pboesu I'll try to do it this Friday, don't hesitate to remind me here.
