### 2021-06-14 #46 @pboesu
@rougier I have just added the metadata for the reviewers and editor and minted a doi for the [code archive](https://doi.org/10.5281/zenodo.4943258) . Let me know if you need me to do anything else.
