### 2021-06-14 #46 @rougier
Acutally we'e now using software heritage for the code. If you go here: https://www.softwareheritage.org/save-and-reference-research-software/ It's quite straightforward to get the SWID.
