### 2021-06-14 #46 @pboesu
OK, I've removed the Zenodo DOI and filled in the swh field in metadata.yaml instead. The snapshot on software heritage is here: https://archive.softwareheritage.org/swh:1:dir:fa45767e4bb56f7807fe2081cfe069881f8b049e
