### 2021-06-16 #46 @rougier
Thanks. The sandboxed version is at https://sandbox.zenodo.org/record/865592. Can you check everything's ok?
