### 2021-06-21 #46 @rougier
It's online at https://zenodo.org/record/5006005 and will soon appear on ReScience website.
