### 2021-06-21 efdfe64
Added publication information
