+++ article/metadata.tex
@@ -2,25 +2,25 @@
 
 \def \codeURL{https://github.com/pboesu/replication_candy_1991}
 \def \codeDOI{}
-\def \codeSWH{}
+\def \codeSWH{swh:1:dir:fa45767e4bb56f7807fe2081cfe069881f8b049e}
 \def \dataURL{https://github.com/pboesu/replication_candy_1991}
 \def \dataDOI{}
-\def \editorNAME{}
-\def \editorORCID{}
-\def \reviewerINAME{}
-\def \reviewerIORCID{}
-\def \reviewerIINAME{}
-\def \reviewerIIORCID{}
-\def \dateRECEIVED{}
-\def \dateACCEPTED{}
-\def \datePUBLISHED{}
+\def \editorNAME{Timothée Poisot}
+\def \editorORCID{0000-0002-0735-5184}
+\def \reviewerINAME{Juniper L. Simonis}
+\def \reviewerIORCID{0000-0001-9798-0460}
+\def \reviewerIINAME{A. Andrew. M. MacDonald}
+\def \reviewerIIORCID{0000-0003-1162-169X}
+\def \dateRECEIVED{22 September 2020}
+\def \dateACCEPTED{03 June 2021}
+\def \datePUBLISHED{16 June 2021}
 \def \articleTITLE{[Re] Modeling Insect Phenology Using Ordinal Regression and Continuation Ratio Models}
 \def \articleTYPE{Replication}
 \def \articleDOMAIN{Ecology}
 \def \articleBIBLIOGRAPHY{bibliography.bib}
-\def \articleYEAR{2020}
+\def \articleYEAR{2021}
 \def \reviewURL{}
-\def \articleABSTRACT{abstract text goes here}
+\def \articleABSTRACT{This article reimplements a number of ordinal regression models in R and succesfully replicates their application to insect phenology data.}
 \def \replicationCITE{Brian Dennis, William P. Kemp, Roy C. Beckwith, Stochastic Model of Insect Phenology: Estimation and Testing, Environmental Entomology, Volume 15, Issue 3, 1 June 1986, Pages 540–546}
 \def \replicationBIB{dennis1986stochastic}
 \def \replicationURL{https://doi.org/10.1093/ee/15.3.540}
@@ -29,10 +29,10 @@
 \def \contactEMAIL{pboesu@gmail.com}
 \def \articleKEYWORDS{Insects, categorical regression, R, sequential model}
 \def \journalNAME{ReScience C}
-\def \journalVOLUME{4}
+\def \journalVOLUME{7}
 \def \journalISSUE{1}
-\def \articleNUMBER{}
-\def \articleDOI{}
+\def \articleNUMBER{5}
+\def \articleDOI{10.5281/zenodo.5006005}
 \def \authorsFULL{Philipp H. Boersch-Supan}
 \def \authorsABBRV{P.H. Boersch-Supan}
 \def \authorsSHORT{Boersch-Supan}
