+++ article/article.tex
@@ -6,7 +6,7 @@
 \begin{document}
 
 \input{header.tex}
-\linenumbers
+% \linenumbers
 \input{content.tex}
 
 \hypersetup{linkcolor=black,urlcolor=darkgray}
